% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.sim.constant.R
\name{bd.sim.constant}
\alias{bd.sim.constant}
\title{Constant rate Birth-Death simulation}
\usage{
bd.sim.constant(
  n0,
  lambda,
  mu,
  tMax,
  nFinal = c(0, Inf),
  nExtant = c(0, Inf),
  trueExt = FALSE
)
}
\arguments{
\item{n0}{Initial number of species. Usually 1, in which case the simulation
will describe the full diversification of a monophyletic lineage. Note that
when \code{lambda} is less than or equal to \code{mu},  many simulations 
will go extinct before speciating even once. One way of generating large 
sample sizes in this case is to increase \code{n0}, which will simulate the
diversification of a paraphyletic group.}

\item{lambda}{Speciation rate. Must be constant, and greater than or equal
to zero.}

\item{mu}{Extinction rate, similar to above.}

\item{tMax}{Ending time of simulation, in million years after the clade 
origin. Any species still living after \code{tMax} is considered extant, and 
any species that would be generated after \code{tMax} is not present in the
return.}

\item{nFinal}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of species at the end of the simulation. Default value is
\code{c(0, Inf)}, so that any number of species (including zero, the
extinction of the whole clade) is accepted. If different from default value,
simulation will restart until the number of total species at \code{tMax}
is in the \code{nFinal} interval.}

\item{nExtant}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of extant species at the end of the simulation. Equal to
\code{nFinal} in every respect except for that.

Note: The function returns \code{NA} if it runs for more than \code{100000}
iterations without fulfilling the requirements of \code{nFinal} and
\code{nExtant}.

Note: Using values other than the default for \code{nFinal} and
\code{nExtant} will condition simulation results.}

\item{trueExt}{A \code{logical} indicating whether the function should return
true or truncated extinction times. When \code{TRUE}, time of extinction of 
extant species will be the true time, otherwise it will be \code{NA} if a 
species is alive at the end of the simulation.}
}
\value{
A \code{sim} object, containing extinction times, speciation times,
parent, and status information for each species in the simulation. See 
\code{?sim}.
}
\description{
Simulates a species birth-death process with constant rates for any number of
starting species. Allows for constraining results on the number of species at 
the end of the simulation, either total or extant. Returns a \code{sim} object 
(see \code{?sim}). It may return true extinction times or simply information on
whether species lived after the maximum simulation time, depending on input. 
For time-varying and age-dependent simulations, see \code{bd.sim.general}. For a
function that unites both scenarios, see \code{bd.sim}.
Please note while time runs from \code{0} to \code{tMax} in the simulation, it 
returns speciation/extinction times as \code{tMax} (origin of the group) to 
\code{0} (the "present" and end of simulation), so as to conform to other
packages in the literature.
}
\examples{

###
# first, let us try no extinction

# initial number of species
n0 <- 1

# maximum simulation time
tMax <- 40

# speciation
lambda <- 0.1

# extinction
mu <- 0

# run the simulation
sim <- bd.sim.constant(n0, lambda, mu, tMax, nFinal = c(2, Inf))

# we can plot the phylogeny to take a look
if (requireNamespace("ape", quietly = TRUE)) {
  phy <- make.phylo(sim)
  ape::plot.phylo(phy)
}

###
# now let us try to turn extinction up a bit

# initial number of species
n0 <- 1

# maximum simulation time
tMax <- 40

# speciation
lambda <- 0.1

# extinction
mu <- 0.04

# run the simulation
sim <- bd.sim.constant(n0, lambda, mu, tMax, nFinal = c(2, Inf))

# we can plot the phylogeny to take a look
if (requireNamespace("ape", quietly = TRUE)) {
  phy <- make.phylo(sim)
  ape::plot.phylo(phy)
}

###
# we can also try a pure-death process

# initial number of species - note the high number, so we get an appreciable
# sample size
n0 <- 100

# maximum simulation time
tMax <- 40

# speciation
lambda <- 0

# extinction
mu <- 0.02

# run the simulation
sim <- bd.sim.constant(n0, lambda, mu, tMax)

# of course in this case there are no phylogenies to plot

###
# note nFinal has to be sensible
\dontrun{
# this would return a warning, since it is virtually impossible to get 100
# species at a process with diversification rate -0.09 starting at n0 = 1
sim <- bd.sim.constant(1, lambda = 0.01, mu = 1, tMax = 100, 
                      nFinal = c(100, Inf))
}

}
\author{
Bruno do Rosario Petrucci.
}
