% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.sim.general.R
\name{bd.sim.general}
\alias{bd.sim.general}
\title{Non-constant rate Birth-Death simulation}
\usage{
bd.sim.general(
  n0,
  lambda,
  mu,
  tMax,
  lShape = NULL,
  mShape = NULL,
  nFinal = c(0, Inf),
  nExtant = c(0, Inf),
  trueExt = FALSE
)
}
\arguments{
\item{n0}{Initial number of species. Usually 1, in which case the simulation
will describe the full diversification of a monophyletic lineage. Note that
when \code{lambda} is less than or equal to \code{mu},  many simulations 
will go extinct before speciating even once. One way of generating large 
sample sizes in this case is to increase \code{n0}, which will simulate the
diversification of a paraphyletic group.}

\item{lambda}{Function that defines the speciation rate over time. It will 
either be interpreted the rate of an exponential distribution, or the scale 
of a Weibull distribution if \code{lShape != NULL}. Can be constant, to 
allow for mixing of constant and non-constant rates. One can use constructs 
such as \code{ifelse()} to create rates whose underlying model change over 
time (see the last examples). Note that \code{lambda} should always be 
greater than or equal to zero.}

\item{mu}{Similar to above, but for the extinction rate.

Note: rates should be considered as running from \code{0} to \code{tMax}, as
the simulation runs in that direction even though the function inverts 
speciation and extinction times before returning.

Note: this function is meant to be called by \code{bd.sim}, so it neither
allows for as much flexibility, nor calls \code{make.rate}. If the user 
wishes to use \code{bd.sim.general} with environmental or step-function 
rates, they can generate the rate with \code{make.rate} and supply it to the
function.}

\item{tMax}{Ending time of simulation, in million years after the clade 
origin. Any species still living after \code{tMax} is considered extant, and 
any species that would be generated after \code{tMax} is not present in the
return.}

\item{lShape}{Shape parameter defining the degree of age-dependency in 
speciation rate. This will be equal to the shape parameter in a Weibull 
distribution: as a species' longevity increases (negative age-dependency). 
When larger than one, speciation rate will increase as a species' longevity 
increases (positive age-dependency). It may be a function of time, but 
see note below for caveats therein. Default is \code{NULL}, equivalent to 
an age-independent process. For \code{lShape != NULL} (including when equal 
to one), \code{lambda} will be considered a scale (= 1/rate), and 
\code{rexp.var} will draw a Weibull distribution instead of an exponential. 
This means Weibull(rate, 1) = Exponential(1/rate). Note that even when 
\code{lShape != NULL}, \code{lambda} may still be time-dependent.}

\item{mShape}{Similar to \code{lShape}, but for the extinction rate.

Note: Simulations with time-varying shape behave within theoretical 
expectations for most cases, but if shape is lower than 1 and varies too 
much (e.g. \code{0.5 + 0.5*t}), it can be biased for higher waiting times 
due to computational error. A degree of time dependence of the order of 
0.01 events/my^2 are advisable. It might, although rarely, exhibit a small 
bias when using shape functions with abrupt time variationsIn both cases,
error is still quite low for the purposes of the package.

Note: Shape must be greater than 0. We arbitrarily chose 0.01 as the minimum
accepted value, so if shape is under 0.01 for any reasonable time in the 
simulation, it returns an error.}

\item{nFinal}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of species at the end of the simulation. Default value is
\code{c(0, Inf)}, so that any number of species (including zero, the
extinction of the whole clade) is accepted. If different from default value,
simulation will restart until the number of total species at \code{tMax}
is in the \code{nFinal} interval. Note that \code{nFinal} must be a sensible
vector. The function will error if its maximum is lower than \code{1}, or if
its length is not \code{2}.}

\item{nExtant}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of extant species at the end of the simulation. Equal to
\code{nFinal} in every respect except for that.

Note: The function returns \code{NA} if it runs for more than \code{100000}
iterations without fulfilling the requirements of \code{nFinal} and
\code{nExtant}.

Note: Using values other than the default for \code{nFinal} and
\code{nExtant} will condition simulation results.}

\item{trueExt}{A \code{logical} indicating whether the function should return
true or truncated extinction times. When \code{TRUE}, time of extinction of 
extant species will be the true time, otherwise it will be \code{NA} if a 
species is alive at the end of the simulation.

Note: This is interesting to use to test age-dependent extinction. 
Age-dependent speciation would require all speciation times (including
the ones after extinction) to be recorded, so we do not attempt to add an
option to account for that. Since age-dependent extinction and speciation
use the same underlying process, however, if one is tested to satisfaction
the other should also be in expectations.}
}
\value{
A \code{sim} object, containing extinction times, speciation times,
parent, and status information for each species in the simulation. See 
\code{?sim}.
}
\description{
Simulates a species birth-death process with general rates for any number of
starting species. Allows for the speciation/extinction rate to be (1) a 
constant, or (2) a function of time. Allows for constraining results on the 
number of species at the end of the simulation, either total or extant. The 
function can also take an optional shape argument to generate age-dependence
on speciation and/or extinction, assuming a Weibull distribution as a model 
of age-dependence. Returns a \code{sim} object (see \code{?sim}). It may 
return calculated extinction times or simply information on whether species 
lived after the maximum simulation time, depending on input. For constant 
rate simulations, see \code{bd.sim.constant}. For a function that unites all
scenarios, see \code{bd.sim}. \code{bd.sim} also allows for extra inputs, 
creating a time-dependent only rate internally through \code{make.rate}. For
similar flexibility, use \code{make.rate} to generate the desired rate.
Please note while time runs from \code{0} to \code{tMax} in the simulation,
it returns speciation/extinction times as \code{tMax} (origin of the group) 
to \code{0} (the "present" and end of simulation), so as to conform to other
packages in the literature.
}
\author{
Bruno do Rosario Petrucci.
}
\keyword{Internal}
