% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.sim.quasse.R
\name{bd.sim.quasse}
\alias{bd.sim.quasse}
\title{QuaSSE simulation}
\usage{
bd.sim.quasse(
  n0,
  lambda,
  mu,
  condition = "time",
  tMax = Inf,
  N = Inf,
  nTraits = 1,
  nFocus = 1,
  X0 = 0,
  sigma = list(1),
  drift = 0,
  bounds = NULL,
  nFinal = c(0, Inf),
  nExtant = c(0, Inf)
)
}
\arguments{
\item{n0}{Initial number of species. Usually 1, in which case the simulation
will describe the full diversification of a monophyletic lineage. Note that
when \code{lambda} is less than or equal to \code{mu},  many simulations 
will go extinct before speciating even once. One way of generating large 
sample sizes in this case is to increase \code{n0}, which will simulate the
diversification of a paraphyletic group.}

\item{lambda}{Function to hold the speciation rate over time. It should
either be a constant, or a function of one argument. For each species a 
trait evolution simulation will be run, and then used to calculate the final 
speciation rate. Note that \code{lambda} should always be greater than or 
equal to zero.}

\item{mu}{Similar to above, but for the extinction rate.}

\item{condition}{Whether to condition tree by total simulation running time,
or the number of species alive at the end of the simulation. If set to 
\code{"time"}, simulation will run for \code{tMax} million years. If set
to \code{"number"}, simulation will run until there are \code{N} species
alive at a given time. 

Note: this manner of conditioning for a number of tips conditions the 
phylogeny to have shorter terminal branches. While this might not be an 
issue for low values of extinction fraction \code{mu/lambda}, future 
implementation of the correct conditioning is planned.}

\item{tMax}{Ending time of simulation, in million years after the clade 
origin. Any species still living after \code{tMax} is considered extant, and 
any species that would be generated after \code{tMax} is not present in the
return.}

\item{N}{Number of species at the end of the simulation, if \code{condition}
equals \code{"number"}. End of the simulation will be set for the first time
a species alive at a period where \code{N} species are alive would go 
extinct.}

\item{nTraits}{The number of traits to be considered. \code{lambda} and 
\code{mu} need not reference every trait simulated.}

\item{nFocus}{Trait of focus, i.e. the one that rates depend on. If it is 
one number, that will be the trait of focus for both speciation and 
extinction rates. If it is of length 2, the first will be the focus for
the former, the second for the latter.}

\item{X0}{Initial trait value for original species. Can be a constant or a 
vector of length \code{nTraits}.}

\item{sigma}{Brownian motion parameter. The variance at time \code{t} after
the start of the simulation will be \code{sigma^2*t}}

\item{nFinal}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of species at the end of the simulation. Default value is
\code{c(0, Inf)}, so that any number of species (including zero, the
extinction of the whole clade) is accepted. If different from default value,
simulation will restart until the number of total species at \code{tMax}
is in the \code{nFinal} interval. Note that \code{nFinal} must be a sensible
vector. The function will error if its maximum is lower than \code{1}, or if
its length is not \code{2}.}

\item{nExtant}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of extant species at the end of the simulation. Equal to
\code{nFinal} in every respect except for that.

Note: The function returns \code{NA} if it runs for more than \code{100000}
iterations without fulfilling the requirements of \code{nFinal} and
\code{nExtant}.

Note: Using values other than the default for \code{nFinal} and
\code{nExtant} will condition simulation results.}
}
\value{
A \code{sim} object, containing extinction times, speciation times,
parent, and status information for each species in the simulation, and a 
list object with the trait functions describing the trait value for each
species at each time.
}
\description{
Simulates a species birth-death process following the Quantitative State 
Speciation and Extinction (QuaSSE) model for any number of starting species. 
Allows for the speciation/extinction rate to be (1) a constant, or (2) a 
function of trait values. Traits are simulated to evolve under a Brownian
motion model (see references). Allows for constraining results on the 
number of species at the end of the simulation, either total or extant, 
using rejection sampling. Returns a \code{sim} object (see \code{?sim}), and
a list of data frames describing trait values for each interval. It may
return true extinction times or simply information on whether species lived
after the maximum simulation time, depending on input.
}
\details{
Please note while time runs from \code{0} to \code{tMax} in the simulation, 
it returns speciation/extinction times as \code{tMax} (origin of the group) 
to \code{0} (the "present" and end of simulation), so as to conform to other
packages in the literature.
}
\examples{

###

}
\references{
Maddison W.P., Midford P.E., Otto S.P. 2007. Estimating a binary character’s 
effect on speciation and extinction. Systematic Biology. 56(5):701.

FitzJohn R.G. 2012. Diversitree: Comparative Phylogenetic Analyses of 
Diversification in R. Methods in Ecology and Evolution. 3:1084–1092.
}
\author{
Bruno do Rosario Petrucci.
}
