% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.sim.traits.R
\name{bd.sim.traits}
\alias{bd.sim.traits}
\title{Trait-dependent Birth-Death simulation}
\usage{
bd.sim.traits(
  n0,
  lambda,
  mu,
  tMax,
  lShape = NULL,
  mShape = NULL,
  nTraits = 1,
  traitModel = "BM",
  pars = list(BM = list(sigma2 = 1, X0 = 0)),
  nFinal = c(0, Inf),
  nExtant = c(0, Inf),
  trueExt = FALSE
)
}
\arguments{
\item{n0}{Initial number of species. Usually 1, in which case the simulation 
will describe the full diversification of a monophyletic lineage. Note that
when \code{lambda} is less than or equal to \code{mu},  many simulations will
go extinct before speciating even once. One way of generating large sample
sizes in this case is to increase \code{n0}, which will simulate the
diversification of a paraphyletic group.}

\item{lambda}{Function to hold the speciation rate over time. It will either be
interpreted as an exponential rate, or a Weibull scale if 
\code{lShape != NULL}. Can be constant, to allow for mixing of constant and
non-constant rates. Can also be a function of a list of traits, 
\code{lambda(t, traits)}. For each species a trait evolution simulation will be
run for each trait, and then used to calculate the final speciation rate.
Note that \code{lambda} should always be greater than or equal to zero.}

\item{mu}{Similar to above, but for the extinction rate.

Note: rates should be considered as running from \code{0} to \code{tMax}, as
the simulation runs in that direction even though the function inverts times
before returning in the end.

Note: this function is meant to be called by \code{bd.sim}, so it neither
allows for as much flexibility, nor calls \code{make.rate}. If the user wishes
to use \code{bd.sim.traits} with environmental or step-function rates, they
can generate the rate with \code{make.rate} and supply it to the function.}

\item{tMax}{Ending time of simulation, in million years after the clade origin. 
Any species still living after \code{tMax} is considered extant, and any 
species that would be generated after \code{tMax} is not born.}

\item{lShape}{Shape of the age-dependency in speciation rate. This will be 
equal to the shape parameter in a Weibull distribution: when smaller than one, 
speciation rate will decrease along each species' age (negative 
age-dependency). When larger than one, speciation rate will increase along each
species's age (positive age-dependency). It may be a function of time, but 
see note below for caveats therein. Default is \code{NULL}, equivalent to 
an age-independent process. For \code{lShape != NULL} (including when equal to 
one), \code{lambda} will be considered a scale (= 1/rate), and \code{rexp.var} 
will draw a Weibull distribution instead of an exponential. This means 
Weibull(rate, 1) = Exponential(1/rate). Note that even when 
\code{lShape != NULL}, \code{lambda} may still be time-dependent.}

\item{mShape}{Similar to \code{lShape}, but for the extinction rate.

Note: Time-varying shape is within expectations for most cases, but if it is
lower than 1 and varies too much (e.g. \code{0.5 + 0.5*t}), it can be biased
for higher waiting times due to computational error. Slopes (or equivalent,
since it can be any function of time) of the order of 0.01 are advisable.
It rarely also displays small biases for abrupt variations. In both cases,
error is still quite low for the purposes of the package.

Note: Shape must be greater than 0. We arbitrarily chose 0.01 as the minimum
accepted value, so if shape is under 0.01 for any reasonable time in the 
simulation, it returns an error.}

\item{nTraits}{The number of traits to be considered. \code{lambda} and 
\code{mu} need not reference every trait simulated.}

\item{traitModel}{The model under which each trait is to be simulated. It can
be "BM" for Brownian Motion, "OU" for the Ornstein-Uhlenbeck process, "EB" for
the Early Burst process, or "ST" for discrete state trait evolution (MuSSE). 
If all traits are supposed to follow the same model, can be just one string.}

\item{pars}{A named list of trait parameters. Should be organized as follows. 

\itemize{

\item \code{pars[["BM"]]} should be a named list containing "sigma2" and "X0"
fields. See \code{?traits.bm} for more information.

\item \code{pars[["OU"]]} should be a named list containing "sigma2", "X0",
"mean" and "theta" fields. See \code{?traits.ou} for more information.

\item \code{pars[["EB"]]} should be a named list containing "sigma2", "X0" and
"b" fields. See \code{?traits.eb} for more information.

\item \code{pars[["ST"]]} should be a named list containing "states", "X0" and
"Q" fields. Since \code{X0} and \code{Q} are always of length greater than 1,
they must be passed inside a list to allow for tests to run correctly. See 
\code{?traits.states} for more information.
}}

\item{nFinal}{A \code{vector} of length \code{2}, indicating an interval of 
acceptable number of species at the end of the simulation. Default value is 
\code{c(0, Inf)}, so that any number of species (including zero, the extinction
of the whole clade) is accepted. If different from default value, the process
will run until the number of total species reaches a number in the interval
\code{nFinal}.}

\item{nExtant}{A \code{vector} of length \code{2}, indicating an interval of
acceptable number of extant species at the end of the simulation. Equal to 
\code{nFinal} in every respect except for that.

Note: The function returns \code{NA} if it runs for more than \code{100000}
iterations without fulfilling the requirements of \code{nFinal} and 
\code{nExtant}.}

\item{trueExt}{A \code{logical} indicating whether the function should return
true or truncated extinction times. When \code{TRUE}, time of extinction of 
extant species will be the true time, otherwise it will be \code{NA} if a 
species is alive at the end of the simulation.

Note: This is interesting to use to test age-dependent extinction. 
Age-dependent speciation would require all speciation times (including
the ones after extinction) to be recorded, so we do not attempt to add an
option to account for that. Since age-dependent extinction and speciation
use the same underlying process, however, if one is tested to satisfacton
the other should also be in expectations.}
}
\value{
A \code{sim} object, containing extinction times, speciation times,
parent, and status information for each species in the simulation. See 
\code{?sim}, and a list object with the trait value functions for each
species in the simulation.
}
\description{
Simulates a species birth-death process with time and trait dependent rates
for any number of starting species. Allows for the speciation/extinction rate
to be (1) a constant, or (2) a function of time and a list of trait values.
Traits are simulated to evolve under Brownian Motion, Ornstein-Uhlenbeck,
Early Burst or MuSSE models (see references). 
Allows for constraining results on the number of species at the end of the 
simulation, either total or extant. The function can also take an optional 
shape argument to generate age-dependence on speciation and/or extinction, 
assuming a Weibull distribution as a model of age-dependence. Returns a 
\code{sim} object (see \code{?sim}). It may return true extinction times or 
simply information on whether species lived after the maximum simulation time,
depending on input. For constant rate simulations, see \code{bd.sim.constant}. 
For simulations not involving trait dependence, see \code{bd.sim.general}. For
a function that unites all scenarios, see \code{bd.sim}. \code{bd.sim} also 
allows for extra inputs. For similar flexibility, use \code{make.rate} to 
generate the desired rate and then modify it to include trait dependence.
Please note while time runs from \code{0} to \code{tMax} in the simulation, it
returns speciation/extinction times as \code{tMax} (origin of the group) to 
\code{0} (the "present" and end of simulation), so as to conform to other
packages in the literature.
}
\examples{

###

}
\references{
Felsenstein, J. (1973). Maximum-Likelihood estimation of evolutionary trees
from continuous chracters. American Journal of Human Genetics. 25(5):471-492.

Uhlenbeck, G. E., Ornstein, L. S. (1930). On the theory of Brownian Motion. 
Phys. Rev. 36 (5): 823–841.

Felsenstein, J. (1988). Phylogenies and Quantitative Characters. Annual Review
of Ecology and Systematics. 19:445-71.

Harmon, L. J., Losos, J. B., Davies, T. D., Gillespie, R. G., Gittleman, J. L.,
Jennings, W. B., Hozak, K. H., McPeek, M. A., Moreno-Roark, F., Near, T. J.,
Purvis, A., Ricklefs, R. E., Schluter, D., Schulte II, J. A., Seehausen, O.,
Sidlauskas, B. L., Torres-Carvajal, O., Weir, J. T., Mooers, A. Ø. (2010).
Early Bursts of Body Size and Shape Evolution are Rare in Comparative Data.
Evolution 64-8: 2385-2396.

Maddison W.P., Midford P.E., Otto S.P. 2007. Estimating a binary character’s 
effect on speciation and extinction. Systematic Biology. 56(5):701.

FitzJohn R.G. 2012. Diversitree: Comparative Phylogenetic Analyses of 
Diversification in R. Methods in Ecology and Evolution. 3:1084–1092.
}
\author{
Bruno do Rosario Petrucci.
}
