% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.occurrences.R
\name{bin.occurrences}
\alias{bin.occurrences}
\title{Bin true occurrences into geologic intervals}
\usage{
bin.occurrences(fossils, bins)
}
\arguments{
\item{fossils}{A \code{data.frame} exactly as returned 
by sample.clade(..., returnTrue = FALSE). See \code{?sample.clade)} 
for details.}

\item{bins}{A vector of time intervals corresponding to geological time 
ranges.}
}
\value{
A \code{data.frame} exactly as returned by
\code{sample.clade(..., returnTrue = TRUE)}. See \code{?sample.clade)} for
details.
}
\description{
Given the output of \code{sample.clade(..., returnTrue = FALSE)}, returns 
the occurrence counts in each bin (i.e., the same as 
\code{sample.clade(..., returnTrue = TRUE)}). This helps to trace 
perfect parallels between both output formats of \code{sample.clade}.
}
\details{
This function helps a user bin "true occurrences" directly into
 binned occurrences, allowing for comparisons among "perfectly known" 
 fossil records and records that have a certain resolution (given by the 
 \code{bins} parameter).
}
\examples{

###
# set seed
set.seed(1) 

# run a birth-death simulation
sim <- bd.sim(n0 = 1, lambda = 0.1, mu = 0.05, tMax = 50)

# choose bins
bins <- seq(0, 50, by = 1)

# generate "true" fossil occurrences
fossils_true <- sample.clade(sim, rho = 1, tMax = 50, returnTrue = TRUE)

# bin the true occurrences
fossils_binned <- bin.occurrences(fossils_true, bins)

# compare
fossils_true
fossils_binned

}
\author{
Matheus Januario.
}
