% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{binner}
\alias{binner}
\title{Bin occurrences}
\usage{
binner(x, bins)
}
\arguments{
\item{x}{the list containing occurrence times for a given species.}

\item{bins}{a vector of time intervals corresponding to geological time ranges.}
}
\value{
a list of upper bounds for the occurrence times of a species.
}
\description{
\code{binner} takes the time interval vector in question and the list of
occurrences for a species.
}
\examples{

# this function is simple to test: let us just create an artificial bin and
# occurrence lists and then check it bounds them correctly
x1 <- c(5.2, 4.9, 4.1, 3.2, 1, 0.2)
bins1 <- c(6, 5, 4, 3, 2, 1, 0)
binneredSamp1 <- binner(x1, bins1) # should be (1, 2, 1, 0, 1, 1)

# it should work with any type of number in IntVec
x2 <- c(6.7, 5.03, 4.2, 3.4, 1.2, 0.4)
bins2 <- c(7, bins1)
binneredSamp2 <- binner(x2, bins2) # should be (1, 1, 1, 1, 0, 1, 1)

# let us try with a real simulated species fossil record
sim <- BDSim(1, 0.1, 0.09, 15)
sampled1 <- Sample(1, sim$TE, sim$TS, 1, 15)
bins3 <- c(15.1, 12.3, 10, 7.1, 5.8, 3.4, 2.2, 0)
binneredSample <- binner(sampled1, bins3)
# one can then check by inspection that the binner is indeed counting correctly

}
\author{
written by Matheus Januario and Bruno do Rosario Petrucci
}
