% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2}
\alias{co2}
\title{Jurassic CO2 data}
\format{
A data frame with 53 rows and 2 variables:
\describe{
  \item{t}{A numeric vector representing time since the beginning of the data 
  frame age, 520 million years ago, in million years. We set this from past to 
  present as opposed to present to past since birth-death functions in 
  \code{paleobuddy} consider time going in said direction.}
  \item{co2}{A numeric vector representing CO2 concentration as the ratio of
  CO2 mass at \code{t} over the present.}
}
}
\source{
\url{https://github.com/hmorlon/PANDAPANDA[m}
}
\usage{
data(co2)
}
\description{
CO2 data during the Jurassic. Modified from the \code{co2} set in 
\href{https://github.com/hmorlon/PANDAPANDA[m}{RPANDA}, originally taken from Mayhew 
et al (2008, 2012).
}
\references{
Morlon H. et al (2016) RPANDA: an R package for macroevolutionary analyses on 
phylogenetic trees. \emph{Methods in Ecology and Evolution} 7: 589-597.

Mayhew, P.J. et al (2008) A long-term association between global temperature
and biodiversity, origination and extinction in the fossil record 
\emph{Proc. of the Royal Soc. B} 275:47-53.

Mayhew, P.J. et al (2012) Biodiversity tracks temperature over time 
\emph{Proc. of the Nat. Ac. of Sci. of the USA} 109:15141-15145.

Berner R.A. & Kothavala, Z. (2001) GEOCARB III: A revised model of atmospheric 
CO2 over Phanerozoic time \emph{Am. J. Sci.} 301:182–204.
}
\keyword{datasets}
