% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_utilities.R
\name{extract.clades}
\alias{extract.clades}
\title{Get clades from buddPhylos and phylos}
\usage{
extract.clades(tree, considerSAs = FALSE, considerOrientation = FALSE)
}
\arguments{
\item{tree}{Object of class "buddPhylo" or "phylo".}

\item{considerSAs}{Logical indicating whether sampled ancestors should be
considered in the definition of a clade.}

\item{considerOrientation}{Logical indicating whether speciation node 
orientation should be considered in the definition of a clade.}
}
\value{
A list of length equal to the number of internal nodes, with each
element being a string with taxa in a given clade separated by commas.
If \code{considerSAs = TRUE}, clades defined by an SA rather than speciation
node will end with \code{"\001SA:X"}, where X is the taxon in question.
If \code{considerOrientation = TRUE}, clades defined by a speciation node
will end with \code{"\001ANC:X;Y;..."}, where X, Y, etc. are the taxa in the
ancestral branch.
}
\description{
Get a list of monophyletic clades from a phylogenetic tree. Could be budding
or bifurcating.
}
\author{
Bruno do Rosario Petrucci
}
