% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.lineages.R
\name{find.lineage}
\alias{find.lineage}
\title{Helper function for find.lineages}
\usage{
find.lineage(sim, s)
}
\arguments{
\item{sim}{a simulation from the \code{bd.sim} function. The
function accept simulations with any number of starting species.}

\item{s}{a species number. The function will return a sim object whose common
ancestor is the species s, and s will be renumbered to species 1 in the result.}
}
\value{
a \code{sim} object with \code{s} as the mother species, including
all descendants of \code{s}. An extra list \code{sim$LIN} is included so the
user can identify the order of species in the return with the order of species
in the original simulation.
}
\description{
\code{find.lineage} takes a simulation object, usually from \code{bd.sim},
and a species number, and returns the simulation object that has that
species as a common ancestor.
}
\author{
written by Bruno Petrucci.
}
