% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_findLineage.R
\name{helper_findLineage}
\alias{helper_findLineage}
\title{Helper function for findLineages}
\usage{
helper_findLineage(sim, s)
}
\arguments{
\item{sim}{a simulation from the \code{BDSim} function. The
function accept simulations with any number of starting species.}

\item{s}{a species number. The function will return a sim object whose common
ancestor is the species s, and s will be renumbered to species 1 in the result.}
}
\value{
a \code{sim} object with \code{s} as the mother species, including
all descendants of \code{s}.

Note: \code{helper_findLineage(sim, s)} will be identical to
\code{findLineages(sim, c(s))}, so we omit examples here.
}
\description{
\code{helper_findLineage} takes a simulation object, usually from \code{BDSim},
and a species number, and returns the simulation object that has that
species as a common ancestor.
}
\author{
written by Bruno Petrucci.
}
