% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.buddPhylo.R
\name{make.buddPhylo}
\alias{make.buddPhylo}
\title{Generate a budding phylogenetic tree from simulations}
\usage{
make.buddPhylo(sim, fossils = NULL, returnTrueExt = TRUE)
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation 
times, parent, and status information for each species in the simulation. 
See \code{?sim}.}

\item{fossils}{A data frame with a \code{"Species"} column and a
\code{SampT} column, usually an output of the \code{sample.clade}
function. Species names must contain only one number each, corresponding
to the order of the \code{sim} vectors. Note that we require it to have a
\code{SampT} column, i.e. fossils must have an exact age. This assumption
might be relaxed in the future.}

\item{returnTrueExt}{A logical indicating whether to include in tree the
tips representing the true extinction time of extinct species. If set to
\code{TRUE} (default), the returned tree will include those tips. If 
\code{FALSE}, they will be dropped and instead the last sampled fossil of
a given species will be the last sampled tip of that species. Note that if
a species was not sampled it will then not appear in the tree. If no fossils
have been added to the tree with the parameter \code{fossils}, this will
have the equivalent effect as the ape function \code{drop.fossil}, returning
an ultrametric tree.}
}
\value{
A \code{buddPhylo} object corresponding to the simulation given. See
\code{?buddPhylo} for details. The \code{lineage} column will be used to 
list the specific occurrence of each taxon--i.e. the first fossil occurrence
of species \code{t1} will be \code{t1.1}, the second \code{t1.2}, etc. If the
species is extant, \code{t1} will be the name of the extant tip.
}
\description{
\code{make.buddPhylo} generates a budding phylogenetic tree
from a birth-death simulation (e.g. an output from \code{bd.sim}) and 
(optionally) a fossil record (e.g. an output from \code{sample.clade}).
}
\examples{

# set seed
set.seed(1)

# run a simple birth-death simulation
sim <- bd.sim(1, 0.2, 0.05, 20)

# and sample fossils on it
sample <- suppressMessages(sample.clade(sim, 0.5, 20))

# build and plot the tree
budd <- make.buddPhylo(sim, sample)
plot(budd, show.SA.label = TRUE)

# by default, we include species that weren't sampled (e.g. t7 in this case)
# we can exclude them (and also exclude the time interval between the last
# sample and extinction of a species through the returnTrueExt argument
budd_all <- make.buddPhylo(sim, sample, returnTrueExt = FALSE)
plot(budd_all, show.SA.label = TRUE)
# note the lines that disappeared, leading some species to be sampled
# ancestors, and others to not be in the tree at all

# of course we could also draw the phylogeny representing the
# birth-death simulation only, before fossils are sampled at all
budd_sim <- make.buddPhylo(sim)
plot(budd_sim)
# this allows you to visualize a simulation in its entirety

}
\author{
Bruno do Rosario Petrucci.
}
