% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePhylo.R
\name{MakePhylo}
\alias{MakePhylo}
\title{Converts a paleobuddy simulation into a phylogeny}
\usage{
MakePhylo(sim)
}
\arguments{
\item{sim}{a simulation from the \code{BDSim} function.}
}
\value{
A \code{phylo} object.
}
\description{
\code{MakePhylo} generates a \code{phylo} object using a simulation from the
\code{BDSim} function. The phylogeny follows a "Hennigian" (sensu Ezard et
al 2011) format. If the simulation has only one lineage, the function
returns \code{NA} as there is no phylogeny for a simulation with only one
lineage.
}
\examples{

#Generating a phylogeny using constant rates
library(ape)
sim<-BDSim(N0 = 1, pp = 0.2, qq = 0.05, tmax = 10)
while(length(sim$TE) < 2){ #in case first simulation has only one species
  sim<-BDSim(N0 = 1, pp = 0.2, qq = 0.05, tmax = 10)
}
phy<-MakePhylo(sim)
plot.phylo(phy)
# we can also plot only the molecular phylogeny
plot.phylo(drop.fossil(phy))

# this works for sim generated with any of the scenarios in \\code{BDSim}, of course
sim<-BDSim(N0=1, pp=function(t) 0.12+0.01*t,qq=10, tmax=10, eshape=1.3)
while(length(sim$TE) < 2){ #in case first simulation has only one species
  sim<-BDSim(N0=1, pp=function(t) 0.12+0.01*t,qq=10, tmax=10, eshape=1.3)
}
phy<-MakePhylo(sim)
plot.phylo(phy)

}
\author{
Function written by Matheus Januario. Reference: Ezard, T. H.,
Pearson, P. N., Aze, T., & Purvis, A. (2012). The meaning of birth and death
(in macroevolutionary birth-death models). Biology letters, 8(1), 139-142.
}
