% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalFileFunctions.R
\name{parse.newick.annotations.notrans}
\alias{parse.newick.annotations.notrans}
\title{Parse and organize annotations from a newick tree}
\usage{
parse.newick.annotations.notrans(newick)
}
\arguments{
\item{newick}{A character containing an annotated tree in newick 
(parenthetic) format (e.g., as output from \code{parse.nexus}).}
}
\value{
A data frame containing node information on lineage (taxon + 
first/last occurrence), taxon (taxonomic name associated with the branch), 
orientation, branch length, node name, and any other annotations 
associated with every node in a tree. Note the function does not return a 
\code{buddPhylo} object.
}
\description{
This function finds and sorts any node-specific annotation in a nexus file.
}
\author{
Matheus Januario
}
