% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalFileFunctions.R
\name{parse.nexus}
\alias{parse.nexus}
\title{Subset and parse a nexus file}
\usage{
parse.nexus(file, treeNumbers = NULL)
}
\arguments{
\item{file}{A character containing a nexus file name.}

\item{treeNumbers}{A numeric vector containing which trees should be read.
If \code{NULL} (default), the function reads all trees within \code{file}.}
}
\value{
A character with the parenthetic format within the nexus file.
If more than one tree is read, a list of characters is returned in the
order given by \code{treeNumbers}
}
\description{
This function reads and parses one or several trees in a NEXUS file.
}
\details{
The does not take note on  whichever name was assigned for any tree
in the nexus file.
}
\author{
Matheus Januario
}
