% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo.R
\name{plot.buddPhylo}
\alias{plot.buddPhylo}
\title{Plot a budding phylogenetic tree}
\usage{
\method{plot}{buddPhylo}(
  x,
  type = "phylogram",
  cex = par("cex"),
  tipCex = par("cex"),
  nodeCex = par("cex"),
  tip.color = par("col"),
  lineageAsNodeLabels = FALSE,
  sampAncJitterMean = 0,
  sampAncJitterSD = 0.3,
  show.tip.label = TRUE,
  show.SA.label = FALSE,
  show.node.label = FALSE,
  edge.color = "black",
  edge.width = NULL,
  edge.lty = NULL,
  node.color = "grey30",
  node.width = NULL,
  node.lty = NULL,
  font = 3,
  adj = NULL,
  srt = 0,
  label.offset = 0,
  underscore = FALSE,
  align.tip.label = FALSE,
  x.lim = NULL,
  y.lim = NULL,
  direction = "rightwards",
  rotate.tree = 0,
  open.angle = 0,
  add.fan.tip.guide = TRUE,
  ...
)
}
\arguments{
\item{x}{Object of class "buddPhylo".}

\item{type}{a character string specifying the type of phylogeny to be drawn;
it must be one of "phylogram" (the default) or "fan".}

\item{cex}{a numeric value giving the factor scaling of the tip and node
labels (character expansion). The default is to take the current value from
the graphical parameters.}

\item{tipCex}{a numeric value giving the factor scaling of the tip
labels (character expansion). The default is to take the
current value from the graphical parameters.}

\item{nodeCex}{a numeric value giving the factor scaling of the node
labels (character expansion). The default is to take the
current value from the graphical parameters.}

\item{tip.color}{the colors used for the tip labels, eventually recycled.}

\item{lineageAsNodeLabels}{A logical indicating whether the information on 
\code{buddPhylo$lineage} must be plotted as node labels, instead of 
\code{buddPhylo$name} (default).}

\item{sampAncJitterMean}{Bias (average) of the jitter associated to each 
sampled ancestor in the tree. The jitter is added to the y-coordinate of the
parent lineage as a normally-distributed random deviation.}

\item{sampAncJitterSD}{Spreading (standard deviation) of the jittering
associated to each sampled ancestor in the tree. The jitter is added to the 
y-coordinate of the parent lineage as a normally-distributed random 
deviation.}

\item{show.tip.label}{A logical indicating whether to show the tip labels on
the phylogeny (defaults to \code{TRUE}, i.e. the labels are shown).}

\item{show.SA.label}{A logical indicating whether to show the sampled 
ancestor labels on the phylogeny (defaults to \code{FALSE}, i.e. the labels 
are not shown).}

\item{show.node.label}{A logical indicating whether to show the node labels
on the phylogeny (defaults to \code{FALSE}, i.e. the labels are not shown).}

\item{edge.color}{A vector of mode character giving the colors used to draw
the branches of the plotted phylogeny. These are taken to be in the same
order than the component edge of the \code{buddPhylo}. If fewer colors are 
given than the number of edges, then the colors are recycled.}

\item{edge.width}{A numeric vector giving the width of the branches of the
plotted phylogeny. These are taken to be in the same order than the
component edge of the \code{buddPhylo}. If fewer widths are given than the
number of edges, then these are recycled.}

\item{edge.lty}{Same as the previous argument but for line types; 1: plain,
2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash.}

\item{node.color}{A vector of mode character giving the colors used to draw
the perpendicular lines associated with each node of the plotted phylogeny.
These are taken to be in the same order than the component node of the
\code{buddPhylo}. If fewer colors are given than the number of nodes, then
the colors are recycled.}

\item{node.width}{Same as the previous argument, but for line widths.}

\item{node.lty}{Same as the previous argument, but for line types; 1: plain,
2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash.}

\item{font}{An integer specifying the type of font for the labels: 1 (plain
text), 2 (bold), 3 (italic, the default), or 4 (bold italic).}

\item{adj}{A numeric specifying the justification of the text strings of
the labels: 0 (left-justification), 0.5 (centering), or 1
(right-justification). The value is set with respect of direction.}

\item{srt}{A numeric giving how much the labels are rotated in degrees
(negative values are allowed resulting in clock-like rotation); the value
has an effect respectively to the value of direction. This option has no 
effect if type = "fan".}

\item{label.offset}{A numeric giving the space between the nodes and the tips
of the phylogeny and their corresponding labels.}

\item{underscore}{A logical specifying whether the underscores in tip
labels should be written as spaces (the default) or left as are (if 
\code{TRUE}).}

\item{align.tip.label}{A logical value or an integer. If \code{TRUE}, the
tips are aligned and dotted lines are drawn between the tips of the tree and 
the labels. If an integer, the tips are aligned and this gives the type of 
the lines (\code{lty}).}

\item{x.lim}{A 2-characters long numeric vector of length one or two giving
the limit(s) of the x-axis. If \code{NULL}, this is computed with respect to
various parameters such as the string lengths of the labels and the branch 
lengths.}

\item{y.lim}{Same as the previous argument, for the y-axis.}

\item{direction}{A character string specifying the direction of the tree.
Four values are possible: "rightwards" (the default), or "leftwards".}

\item{rotate.tree}{For "fan" trees: the rotation of the whole tree in d
egrees (negative values are accepted).}

\item{open.angle}{If type = "fan", the angle in degrees left blank. Use a 
non-zero value if you want to add a time axis after the tree is plotted.}

\item{add.fan.tip.guide}{A logical indicating if a dotted line connecting 
every tip to its label should be added. Has no effect if 
\code{align.tip.label = FALSE} or if \code{type = "phylogram"}.}

\item{...}{Further arguments inherited from graphics::plot()}
}
\description{
Plot a budding phylogenetic tree
}
\details{
\code{plot.buddPhylo} Plots a \code{buddPhylo} objects as a budding
phylogenetic tree. The function was coded to work similarly, thought not 
exactly like, \code{ape::plot.phylo()}.
}
\author{
Matheus Januario and Bruno do Rosario Petrucci.
}
