% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalFileFunctions.R
\name{read.nexus}
\alias{read.nexus}
\title{Read a buddPhylo tree from a file in nexus format}
\usage{
read.nexus(file, treeNumbers = NULL, timeFromRoot = FALSE)
}
\arguments{
\item{file}{A character containing a nexus file name.}

\item{treeNumbers}{A numeric vector containing which trees should be read.
If \code{NULL} (default), the function reads all trees within \code{file}.}

\item{timeFromRoot}{A logical indicating whether time should be expressed as
time since the root (\code{timeFromRoot=TRUE}). If set to \code{FALSE}
(default), time is expressed as millions of years ago (Mya), taking the tip extant tip that is further from the root as a reference for "present".}
}
\value{
A \code{buddPhylo} with the time adjusted accordingly.
}
\description{
This function reads one or several trees in a NEXUS file.
}
\author{
Matheus Januario
}
