% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp.quasse.R, R/rexp.traits.R
\name{rexp.traits}
\alias{rexp.traits}
\title{Trait dependent Exponential waiting times}
\usage{
rexp.traits(n, rate, traits, now = 0, tMax = Inf, fast = FALSE)

rexp.traits(n, rate, traits, now = 0, tMax = Inf, fast = FALSE)
}
\arguments{
\item{n}{The number of waiting times to return. Usually 1, but we allow for
a higher \code{n} to be consistent with the \code{rexp} function.}

\item{rate}{The rate parameter for the exponential distribution. Can be a
constant, or a vector corresponding to the rate for each state of the trait.}

\item{traits}{The trait data frame detailing trait value for each interval 
of time. Its rows should be

\itemize{

\item \code{value} Trait values. These are assumed to be categorical, and 
therefore always in the range \code{(0, nStates - 1)}. So e.g. for two 
states, we have values 0 and 1. 

\item \code{min} The starting time for each trait value. The first row will
always have either 0 or the species' speciation time here. 

\item \code{max} The ending time for each trait value. Note that the 
\code{max} of a row is the \code{min} of the next row. After the max time of
the last row, the species is assumed to maintain the same trait value it had
last.}}

\item{now}{The current time. Needed if one wants to consider only the 
interval starting at the current time time-varying nate. Note this does 
means the waiting time is \code{>= now}, so we also subtract \code{now} from
the result before returning. The default is \code{0}.}

\item{tMax}{The simulation ending time. If the waiting time would be too
high, we return \code{tMax + 0.01} to signify the event never happens, if
\code{fast == TRUE}. Otherwise we return the true waiting time. By default,
\code{tMax} will be \code{Inf}, but if \code{FAST == TRUE} one must supply
a finite value.}

\item{fast}{If set to \code{FALSE}, waiting times larger than \code{tMax} will
not be thrown away. This argument is needed so one can test the function
without bias.}
}
\value{
A vector of waiting times following the exponential distribution 
with the given parameters.

A vector of waiting times following the exponential distribution 
with the given parameters.
}
\description{
Generates a waiting time following an exponential rates varying based on a
trait data frame. Output can be used as the waiting time to an extinction,
speciation, or fossil sampling event. Allows for further customization by 
restricting possible waiting time outputs with arguments for (1) current 
time, to consider only the rates starting at that time, and (2) maximum time
to bound the output and therefore allow for faster calculations if one only
cares about waiting times lower than a given amount.

Generates a waiting time following an exponential rates varying based on a
trait data frame. Output can be used as the waiting time to an extinction,
speciation, or fossil sampling event. Allows for further customization by 
restricting possible waiting time outputs with arguments for (1) current 
time, to consider only the rates starting at that time, and (2) maximum time
to bound the output and therefore allow for faster calculations if one only
cares about waiting times lower than a given amount.
}
\author{
Bruno do Rosario Petrucci.

Bruno do Rosario Petrucci.
}
