% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.age.time.R
\name{sample.age.time}
\alias{sample.age.time}
\title{Time-dependent and age-dependent rate species sampling}
\usage{
sample.age.time(sim, rho, tMax, S = NULL, adFun = NULL, ...)
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation 
times, parent, and status information (extant or extinct) for each species
in the simulation. See \code{?sim}.}

\item{rho}{Average fossil sampling rate (per species per million years) over 
time. It can be a \code{numeric} describing a constant rate or a 
\code{function(t)} describing the variation in sampling over time. For more 
flexibility on sampling, see \code{make.rate} to create more complex rates.
If \code{adFun} is supplied, it will be used to find the number of 
occurrences during the species duration, and a normalized \code{rho*adFun} 
will determine their distribution along the species duration. Note that 
\code{rho} should always be greater than or equal to zero.}

\item{tMax}{The maximum simulation time, used by \code{rexp.var}. A sampling
time greater than \code{tMax} would mean the occurrence is sampled after the
present, so for consistency we require this argument. This is also required
to ensure time follows the correct direction both in the Poisson process and
in the output.}

\item{S}{A vector of species numbers to be sampled. The default is all 
species in \code{sim}. Species not included in \code{S} will not be sampled 
by the function.}

\item{adFun}{A density function representing the age-dependent preservation
model. It must be a density function, and consequently integrate to 1 
(though this condition is not verified by the function). If not provided, a 
uniform distribution will be used by default. The function must also 
have the following properties:

\itemize{

\item Return a vector of preservation densities for each time in a given
vector \code{t} in geological time. 

\item Be parametrized in the absolute geological time associated to 
each moment in age (i.e. age works relative to absolute geological 
time, in Mya - in other words, the convention is TS > 0). The function 
\emph{does not} directly use the lineage's age (which would mean that
TS = 0 for all species whenever they are born). Because of this, it is
assumed to go from \code{tMax} to \code{0}, as opposed to most functions 
in the package.

\item Should be limited between \code{s} (i.e. the lineage's 
speciation/birth) and \code{e} (i.e. the lineage's extinction/death), 
with \code{s} > \code{e}. It is possible to assign parameters in absolute 
geological time (see third example) and relative to age as long as this 
follows the convention of age expressed in absolute geological time (see 
fourth example).

\item Include the arguments \code{t}, \code{s}, \code{e} and \code{sp}. 
The argument sp is used to pass species-specific parameters (see examples),
allowing for \code{dFun} to be species-inhomogeneous.
}}

\item{...}{Additional parameters used by \code{adFun}. See examples.}
}
\value{
A list of vectors of occurrence times for each species in \code{S}.
}
\description{
Generates a vector of occurrence times for a species in a simulation using 
a Poisson process coupled with age-dependent fossil sampling. Allows for the 
Poisson rate to be (1) constant or (2) time-dependent, and additionally 
allows the distribution of occurrences throughout a species age to be any 
function between speciation and extinction time. For more flexibility, see 
\code{make.rate} and \code{sample.clade}. Note that while the simulation 
occurs in forward time, we return (both in birth-death functions and here)
results in backwards time, so that time is inverted using \code{tMax} both 
at the beginning and end of the function.
}
\author{
Matheus Januario
}
\keyword{Internal}
