% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.time.R
\name{sample.time}
\alias{sample.time}
\title{Constant and time-dependent rate species sampling}
\usage{
sample.time(sim, rho, tMax, S = NULL)
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation 
times, parent, and status information (extant or extinct) for each species
in the simulation. See \code{?sim}.}

\item{rho}{Sampling rate (events per species per million years) over time. 
It can be a \code{numeric} describing a constant rate or a \code{function(t)} 
describing the variation in sampling over time. For more flexibility on 
sampling, see \code{make.rate} to create more complex rates. Note that 
\code{rho} should always be greater than or equal to zero.}

\item{tMax}{The maximum simulation time, used by \code{rexp.var}. A sampling
time greater than \code{tMax} would mean the occurrence is sampled after the
present, so for consistency we require this argument. This is also required
to ensure time follows the correct direction both in the Poisson process and
in the output.}

\item{S}{A vector of species numbers to be sampled. The default is all 
species in \code{sim}. Species not included in \code{S} will not be sampled 
by the function.}
}
\value{
A list of vectors of occurrence times for each species in \code{S}.
}
\description{
Generates a vector of occurrence times for species in a simulation using a
Poisson process. Allows for the Poisson rate to be (1) a constant or (2) a 
function of time. For fossil sampling dependent on species age in addition 
to absolute time, see \code{sample.clade}.
}
\details{
Note that while the Poisson process occurs in forward time, we return (both 
in birth-death functions and here) results in backwards time, so that time 
is inverted using \code{tMax} both at the beginning and end of 
\code{sample.time}.
}
\author{
Bruno do Rosario Petrucci and Matheus Januario.
}
\keyword{Internal}
