% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.chars.R
\name{sim.chars}
\alias{sim.chars}
\title{Discrete character simulation}
\usage{
sim.chars(
  sim,
  sample = NULL,
  nTraits = 1,
  nStates = 2,
  X0 = 0,
  Q = function(i) return(matrix(c(0, 0.1, 0.1, 0), 2, 2)),
  species_clock = 1,
  var_min = 0,
  labels = NULL,
  sim_max = 10000,
  sample_fossil = "first",
  extant_sampling = FALSE
)
}
\arguments{
\item{sim}{A simulation object, usually returned by a birth-death simulation 
function like \code{bd.sim}.}

\item{sample}{A data frame with (at least) three columns: \code{Species} (the
species name, usually \code{tN.M} for the \code{M}th sample of the species 
that originated in \code{sim$TS[N]}), \code{Extant} (boolean indicating
whether the species is extant or not), and \code{SampT} (the fossil sampling
time of that species). Usually returned by a fossil sampling simulation
function like \code{sample.clade}. Default is \code{NULL}, in which case the
traits will only be sampled for the extant species in \code{sim}.}

\item{nTraits}{The number of traits to be simulated. It may be a number or
\code{NULL}. If it is a number, that number of traits will be simulated, with
\code{nStates} and \code{X0} expected to be either a single number or a
vector of length \code{nTraits} indicating the number of states for each
trait. If it is \code{NULL}, \code{nStates} is assumed to be a vector where
the nth element represents the desired number of traits with n states. This
is done because simulating a character with n states does not guarantee the
sampled data (at each specified time for each species) will lead to an n
state character. Setting \code{nTraits} and \code{nStates} as described will
guarantee \code{sum(nStates)} characters, each with the desired number of
states. Default is \code{1}.}

\item{nStates}{The number of states for each character. Can be either a
single number, a vector of numbers of length \code{nTraits}, or a vector of
numbers representing the number of characters to simulate with a given number
of states. See the \code{nTraits} description for details. Default is 
\code{2}.}

\item{X0}{The initial state for each character. Can be either a single number
or a vector of numbers of length \code{nTraits}. If \code{nTraits} is
\code{NULL}, \code{X0} should be a single number representing the starting
state for all characters. Default is \code{0}.}

\item{Q}{The transition matrix for a given number of states. It is assumed
to be a function that takes one argument (the number of states) and returns
a square matrix with that many rows and columns. For different states 
\code{i} and \code{j}, the rate at which a species at state \code{i}
transitions to \code{j} is assumed to be \code{Q(n)[i + 1, j + 1]}, where 
\code{n} is the number of states. Default is a 2x2 matrix with symmetrical
transition rates equal to \code{2}. Note that while the standard is to set 
the diagonal values such that rows sum to \code{0}, the diagonal values are 
never used in our algorithm, and we generally set them to \code{0}.}

\item{species_clock}{The rate modified for each species in the simulation.
It can be a number (in which case it will be a global clock), or a vector of
length \code{length(sim$TS)}. Default is \code{1}, in which case the clock
will be defined by \code{Q}. That is the easiest way to set a strict clock.
For a relaxed clock, remember to set \code{Q} considering relative rates,
and \code{species_clock} will then be a multiplier for each species. Note
that since paleobuddy considers a budding speciation model, it is currently
impossible to have different rates at different branches belonging to the
same species. This might be implemented in the future.}

\item{var_min}{The percentage minimum variation for characters along samples.
After a character is simulated, the number of samples with each state up to
the number of states for that character is checked. If \code{nTraits} is 
\code{NULL}, it is assumed that only traits with exactly the given number of
states sampled will be taken, so a trait is rejected unless every state is
represented in at least \code{var_min*100} percent of species. If 
\code{nTraits} is a number, it is assumed that traits simulated with the 
given number of states, but that did not end up with that many states sampled
will be accepted. As such, then any trait with at least two states being
represented at this percentage or more will be accepted. For example, if
\code{var_min == 0} (the default), a 3-state trait will be accepted if either
all states are at all represented (if \code{nTraits == NULL}), or if at least
two states are (if \code{nTraits} is a number). In other words, only
invariant characters are rejected, but \code{nTraits} defines whether to
match the sampled and simulated number of states exactly.}

\item{labels}{The labels for the states on the output data. If \code{NULL}
(default), the states will be \code{0:(nStates - 1)}. If it is a vector, the
states will be changed to the labels, with state \code{0} being 
\code{labels[1]} and so forth. Note that currently it is assumed that one 
would use the same labels for all characters, so if multiple characters are
simulated with differing numbers of states, \code{labels} should be at least
as long as the number of states of the character with the most states. For
example, if a character with 2 states and one with 3 states are simulated,
and \code{labels} is \code{c("A", "B", "C")}, the first character will have
states \code{"A"} and \code{"B"}, while the second will have all three 
states.}

\item{sim_max}{The maximum number of simulations to run. If this number is
reached before we have simulated the requested number of traits (likely due
to \code{min_var} being too strict for the parameters given), the function
returns \code{NULL}. The default value is \code{10000}.}

\item{sample_fossil}{Which fossil to sample for each species' character. 
Acceptable values are \code{"first"}, \code{"last"}, and \code{"random"}.
If set to \code{"first"} or \code{"last"}, the respective occurrence time of
each species will be used to extract the character state for that species.
If set to \code{"random"}, each character will be taken from one of the
occurrence times, to be chosen by sampling all occurrence times at uniform
probability. This is necessary because the state of a character for a given
species may change throughout its lifetime. The default value is 
\code{"first"} for more predictable results, but a \code{"random"} value may
better emulate how morphological paleontology is carried in practice.}

\item{extant_sampling}{Whether to sample characters at the present for extant
species. Since extant samples often take precedence for the scoring of
morphological characters due to their completeness, one might want to sample
characters at the present for extant species regardless of the value of
\code{sample_fossil}. This argument is not necessary if 
\code{sample_fossil == "last"}. The default is \code{FALSE}, meaning extant
species will be sampled with the same procedure as extinct ones (defined by
\code{sample_fossil}), unless they have no fossil samples, in which case
characters must be sampled at the present.}
}
\value{
A data frame with rows representing sampled species in the simulation
(i.e. extinct species with no fossil samples are excluded), and columns
representing each character.
}
\description{
Simulates a discrete character following a continuous-time Markov Chain
for all species in a \code{sim} and (optionally) a \code{sample} object.
The function simulates the character evolution for every species in 
\code{sim}, with each species taking as a starting state the state of its
parent at the time of speciation. For details on the character evolution
simulation algorithm, see \code{?sim.char.species}. Then, each character is
sampled for extant species and for each fossil sample, with options regarding
the state at which time should be assigned to a given species. Options are
also given regarding the labels of states, and rejection sampling for the
resulting data. Finally, the structure of parameters can accommodate
simulations up to a fixed number of characters, or simulations until we reach
a given number of characters sampled with given numbers of states.
}
\examples{

# set seed
set.seed(1)

# run simple simulations
sim <- bd.sim(1, 0.2, 0.05, 10)
sample <- suppressMessages(sample.clade(sim, 0.5, 10))

###
# a simple example using all default parameters

# one character
chars <- sim.chars(sim, sample)
# note that this represents the states at which each species was at their
# first fossil occurrence times (if they were sampled)

###
# we can make things more complicated

# let us simulate 10 traits
nTraits <- 10

# the first 5 will have 2 states, the last 5 will have 4
nStates <- c(rep(2, 5), rep(4, 5))

# we need to create a function for the state matrix
Q <- function(i) {
  res <- matrix(0.05, i, i)
  diag(res) <- 0
  return(res)
}

# run the simulation
chars <- sim.chars(sim, sample, nTraits, nStates, Q = Q)
# note that character 7 only had two states, even though we simulated 4
# this is fine, of course, but there is a way to force the simulation
# of characters with a number of observed states

# nStates is now a vector of how many characters of each state number we want
nStates <- c(0, 5, 0, 5)

# we simulate with nTraits = NULL
chars <- sim.chars(sim, sample, nTraits = NULL, nStates, Q = Q)
# now all characters have the observed number of states equal to the
# simulated number of states
# note that this might be a minor violation of the Mk model

###
# species don't need to all have the same clock, of course

# return nStates to what it was previously
nStates <- c(rep(2, 5), rep(4, 5))

# we can modify Q to be a relative rate matrix
Q <- function(i) {
  res <- matrix(1 / (i - 1), i, i)
  diag(res) <- 0
  return(res)
}

# and draw species clocks from a lognormal
species_clock <- rlnorm(length(sim$TS), 0, 1)

# we will normalize it so that it has a mean of 1
species_clock <- species_clock / mean(species_clock)

# run the simulation
chars <- sim.chars(sim, sample, nTraits, nStates,
                   Q = Q, species_clock = species_clock)
                   
###
# the labels parameter is useful when simulating DNA

# set the labels
labels <- c("A", "C", "G", "T")

# make a Q matrix corresponding to the Jukes-Cantor model
Q <- function(i) {
  res <- matrix(1/3, 4, 4)
  diag(res) <- 0
  return(res)
}

# run the simulation
chars <- sim.chars(sim, sample, nTraits = 10, nStates = 4, Q = Q,
                   labels = labels)

# you might then want to delete the data for the extinct species
chars[which(!sim$EXTANT), ] <- rep("?", 10)
                   
###
# we can simulate under a more complicated molecular evolution model
# let us consider the HKY85 model (Hasegawa, Kishino and Yano 1985)

# let's assume a starting frequency slightly higher for A and C
pi_A <- pi_C <- 0.3
pi_G <- pi_T <- 0.2

# set starting frequency
X0 <- c(0, 0, 0, 1, 1, 2, 2, 2, 3, 3)
# in a real simulation study you would draw this, but we are just
# keeping it simple

# assume transitions happen twice as often as transverions
kappa <- 2

# set the Q matrix
Q <- function(i) return(matrix(c(0, pi_A, kappa * pi_A, pi_A,
                                 pi_C, 0, pi_C, kappa * pi_C,
                                 kappa * pi_G, pi_G, 0, pi_G,
                                 pi_T, kappa * pi_T, pi_T, 0), 4, 4))
# remember in R matrices are set column-wise

# run the simulation
chars <- sim.chars(sim, sample, nTraits = 10, nStates = 4, X0, Q, 
                   labels = labels)
chars[which(!sim$EXTANT), ] <- rep("?", 10)
                   

}
\references{
Jukes, T. H., Cantor, C. R. 1969. Evolution of protein molecules. New York:
Academic Press. 21-132.

Hasegawa, M., Kishino, H., Yano, T. 1985. Dating of the human-ape splitting 
by a molecular clock of mitochondrial DNA. Molecular Evolution. 
22(2):160-174.

Lewis, P.O. 2001. A likelihood approach to estimating phylogeny from 
discrete morphological character data. Systematic Biology. 50(6):913-925.
}
\author{
Bruno do Rosario Petrucci
}
