% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.eb.R
\name{traits.eb}
\alias{traits.eb}
\title{Simulate trait evolution through the Early Burst model}
\usage{
traits.eb(
  tMax,
  tStart = 0,
  nTraits = 1,
  sigma2 = 1,
  b = -0.05,
  X0 = 0,
  nPoints = 100
)
}
\arguments{
\item{tMax}{The only required parameter. Ending time of the simulation, usually
inherited from a birth-death or sampling function.}

\item{tStart}{The starting time of the simulation. Standard Brownian Motion 
starts at \code{0}, so we set the default value as such, but this argument is
required to accurately return a function based on the trait evolution of a
trait for a species born after the beginning of a birth-death simulation. Note
that the trait value for \code{t = 0} is always \code{0}.}

\item{nTraits}{The number of traits to be simulated. Traits are assumed to be
independent from one another. By default, the function simulates one trait.}

\item{sigma2}{The baseline variance of the normal distribution. The total
variance in trait values will be influenced by \code{b}, leading to a
variance of \code{sigma2 * (exp(b * t) - 1) / b}. The default is \code{1}. Can
also be a vector with size equal to \code{nTraits}.}

\item{b}{Exponential rate determining the speed of exponential decay of
trait value variance. Default is \code{-0.05}. Note it can have positive
values if one wishes to model trait variance increasing with time. Can also
be a vector with size equal to \code{nTraits}.}

\item{X0}{Initial trait value. Default is \code{0}. Can also be a vector with
size equal to \code{nTraits}.}

\item{nPoints}{The total number of points in the discretization of the process.
While Brownian Motion is a continuous process, we must simulate it discretely.
As such, the user can provide a number of time points, by default set to 
\code{100}, for which the exact trait values will be calculated. Remaining
values of time will have values predicted by linear interpolation (see 
\code{?approxfun}).}
}
\value{
A named list of functions, where \code{traits.eb$traitN(t)}
corresponds to the trait value of trait \code{N} at time \code{t}.
}
\description{
Generates trait evolution simulations following the Early Burst model,
an extension of Brownian Motion (see \code{traits.bm}) where the corresponding
variance follows an exponential decay (or increase). It is ideal for simulating
adaptive radiations (see References).
}
\examples{
###
# we can start with a simple trait with all default values

# set a maximum simulation time
tMax <- 10

# a starting time
tStart <- 0

# and a number of time points
nPoints <- 100

# which lets us find a vector of time points
times <- seq(tStart, tMax, (tMax - tStart) / nPoints)

# calculate BM function
ebFunc <- traits.eb(tMax)$trait1

# plot it
plot(times, ebFunc(times), type = 'l', main = "Trait values for trait1",
     xlab = "Time (my)", ylab = "Trait value")

###
# a lot of customization is possible - let us try a different sigma2

# set a maximum simulation time
tMax <- 10

# a starting time
tStart <- 0

# and a number of time points
nPoints <- 100

# which lets us find a vector of time points
times <- seq(tStart, tMax, (tMax - tStart) / nPoints)

# try a lower value for sigma2
sigma2 <- 0.5

# calculate BM function
ebFunc <- traits.eb(tMax, sigma2 = sigma2)$trait1

# plot it
plot(times, ebFunc(times), type = 'l', main = "Trait values for trait1",
     xlab = "Time (my)", ylab = "Trait value")

###
# we can also try a different b value

# set a maximum simulation time
tMax <- 10

# a starting time
tStart <- 0

# and a number of time points
nPoints <- 100

# which lets us find a vector of time points
times <- seq(tStart, tMax, (tMax - tStart) / nPoints)

# try a higher sigma2 this time
sigma2 <- 1.5

# and a faster exponential decay
b <- -0.15

# calculate BM function
ebFunc <- traits.eb(tMax, sigma2 = sigma2, b = b)$trait1

# plot it
plot(times, ebFunc(times), type = 'l', main = "Trait values for trait1",
     xlab = "Time (my)", ylab = "Trait value")

###
# we can calculate any number of traits at once

# set a maximum simulation time
tMax <- 10

# a starting time
tStart <- 0

# and a number of time points
nPoints <- 100

# which lets us find a vector of time points
times <- seq(tStart, tMax, (tMax - tStart) / nPoints)

# if sigma2 is a vector, each trait will have 
# a different variance, in the corresponding order
sigma2 <- c(0.5, 0.1, 0.05)

# we can also have different starting points
X0 <- c(0, 0.2, -0.1)

# and different b, but let us just have all the same
b <- -0.025

# set a seed so we can control ylim and legend position
set.seed(1) 

# calculate BM functions
ebFuncs <- traits.eb(tMax, nTraits = 3, sigma2 = sigma2, b = b, X0 = X0)

# plot them
plot(times, ebFuncs[[1]](times), type = 'l', 
     main = "Trait values for traits with sigma2 = 0.5, 0.1 and 0.05",
     xlab = "Time (my)", ylab = "Trait value", ylim = c(-2.5, 3))
lines(times, ebFuncs[[2]](times), col = 'RED')
lines(times, ebFuncs[[3]](times), col = 'BLUE')
legend(x = 1, y = -1, legend = c(0.5, 0.1, 0.05),
       col = c('BLACK', 'RED', 'BLUE'), lty = c(1, 1, 1))
     
}
\references{
Harmon, L. J., Losos, J. B., Davies, T. D., Gillespie, R. G., Gittleman, J. L.,
Jennings, W. B., Hozak, K. H., McPeek, M. A., Moreno-Roark, F., Near, T. J.,
Purvis, A., Ricklefs, R. E., Schluter, D., Schulte II, J. A., Seehausen, O.,
Sidlauskas, B. L., Torres-Carvajal, O., Weir, J. T., Mooers, A. Ø. (2010).
Early Bursts of Body Size and Shape Evolution are Rare in Comparative Data.
Evolution 64-8: 2385-2396.
}
\author{
Bruno do Rosario Petrucci.
}
