% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.summary.R
\name{traits.summary}
\alias{traits.summary}
\title{Summarizing trait data}
\usage{
traits.summary(sim, traits, nFocus = 1, fossils = NULL, selection = "all")
}
\arguments{
\item{sim}{...}

\item{traits}{...}

\item{nFocus}{...}

\item{fossils}{...}

\item{selection}{...}
}
\value{

}
\description{
...
}
\examples{

}
\references{

}
\author{
Bruno do Rosario Petrucci
}
