% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_utilities.R
\name{adjust.timescale.buddPhylo}
\alias{adjust.timescale.buddPhylo}
\title{Adjust the timescale of a buddPhylo pbject}
\usage{
adjust.timescale.buddPhylo(buddPhylo, timeFromRoot = FALSE)
}
\arguments{
\item{buddPhylo}{Object of class "buddPhylo"}

\item{timeFromRoot}{A logical indicating whether time should be expressed as
time since the root (\code{timeFromRoot=TRUE}). If set to \code{FALSE}
(default), time is expressed as millions of years ago (Mya), taking the tip
extant tip that is further from the root as a reference for "present".}
}
\value{
A \code{buddPhylo} with the time adjusted accordingly.
}
\description{
Given a vector of fossil occurrences and time bins to represent geological
ranges, returns the occurrence counts in each bin.
}
\author{
Matheus Januario
}
