% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhyloTreeHandling.R
\name{as.phylo.buddPhylo}
\alias{as.phylo.buddPhylo}
\title{Genertaing a phlyo object from a buddPhylo object}
\usage{
\method{as.phylo}{buddPhylo}(buddPhylo)
}
\arguments{
\item{buddPhylo}{Object of class "buddPhylo"}
}
\value{
A \code{phylo} object compatible with fucntions from package \code{ape}
}
\description{
Generates a phylogeny ina \code{phylo} format based on the ifnromation within 
a \code{buddPhylo} object.
}
\author{
Matheus Januario
}
