% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_utilities.R
\name{as.phylo.buddPhylo}
\alias{as.phylo.buddPhylo}
\title{Generating a phylo object from a buddPhylo object}
\usage{
\method{as.phylo}{buddPhylo}(x, ...)
}
\arguments{
\item{x}{A \code{buddPhylo} object.}

\item{...}{Further arguments inherited from generics.}
}
\value{
A \code{phylo} object compatible with functions from the \code{ape}
package.
}
\description{
Generates a phylogeny in a \code{phylo} format based on the information
within a \code{buddPhylo} object.
}
\author{
Matheus Januario and Bruno do Rosario Petrucci.
}
