% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.sim.constant.R
\name{bd.sim.constant}
\alias{bd.sim.constant}
\title{Constant rate Birth-Death simulation}
\usage{
bd.sim.constant(n0 = 1, pp, qq, tMax, nFinal = c(0, Inf), extOnly = FALSE)
}
\arguments{
\item{n0}{Initial number of species, usually 1. Good parameter
to tweak if one is observing a low sample size when testing.}

\item{pp}{Speciation rate. Must be constant.}

\item{qq}{Extinction rate, similar to above.

Note: \code{pp} and \code{qq} must always be greater than 0}

\item{tMax}{Ending time of simulation. Any species still living
after tMax is considered extant, and any species that would be generated
after \code{tMax} is not born.}

\item{nFinal}{An interval of acceptable number of species at the end of the
simulation. If not supplied, default is \code{c(0, Inf)}, so that any number
of species is accepted. If supplied, \code{bd.sim.constant} will run until the
number of total species generated, or, if \code{extOnly = TRUE}, the number of
extant species at the end of the simulation, lies within the interval.}

\item{extOnly}{A boolean indicating whether \code{nFinal} should be taken as
the number of total or extant species during the simulation. If \code{TRUE},
\code{bd.sim.constant} will run until the number of extant species lies within
the \code{nFinal} interval. If \code{FALSE}, as default, it will run until the
total number of species generated lies within that interval.}
}
\value{
A list of vectors, as follows

\describe{
\item{\code{TE}}{List of extinction times, with -0.01 as the time of
extinction for extant species.}

\item{\code{TS}}{List of speciation times, with tMax+0.01 as the time of
speciation for species that started the simulation.}

\item{\code{PAR}}{List of parents. Species that started the simulation have
NA, while species that were generated during the simulation have their
parent's number. Species are numbered as they are born.}

\item{\code{EXTANT}}{List of booleans representing whether each species is
extant.}}
}
\description{
Simulates a species birth-death process with constant rates for any number
of starting species. Returns an object containing lists of speciation times,
extinction times, parents and status (extant or not). Allows for constraining
on the number of species at the end of the simulation, either total or extant.
Note that while time runs from \code{0} to \code{tmax} on the function itself,
it runs from \code{tmax} to \code{0} on the lists returned to conform with the
literature.
}
\examples{

# we can show a couple scenarios

###
# first, extinction 0

# initial number of species
n0 <- 1

# maximum simulation time
tMax <- 40

# speciation
p <- 0.1

# extinction
q <- 0

# run the simulation
sim <- bd.sim.constant(n0, p, q, tMax, nFinal = c(2, Inf))

# we can plot the phylogeny to take a look
if (requireNamespace("ape", quietly = TRUE)) {
  phy <- make.phylo(sim)
  ape::plot.phylo(phy)
}

###
# now let us try to turn extinction up a bit

# initial number of species
n0 <- 1

# maximum simulation time
tMax <- 40

# speciation
p <- 0.1

# extinction
q <- 0.04

# run the simulation
sim <- bd.sim.constant(n0, p, q, tMax, nFinal = c(2, Inf))

# we can plot the phylogeny to take a look
if (requireNamespace("ape", quietly = TRUE)) {
  phy <- make.phylo(sim)
  ape::plot.phylo(phy)
}

###
# we can also try a pure-death process

# initial number of species - note the high number, so we get an appreciable
# sample size
n0 <- 100

# maximum simulation time
tMax <- 40

# speciation
p <- 0

# extinction
q <- 0.02

# run the simulation
sim <- bd.sim.constant(n0, p, q, tMax)

# of course in this case there are no phylogenies to plot

}
\author{
Bruno do Rosario Petrucci.
}
