% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSimConstant.R
\name{BDSimConstant}
\alias{BDSimConstant}
\title{Constant rate Birth-Death simulation}
\usage{
BDSimConstant(n0 = 1, pp, qq, tMax, nFinal = c(0, Inf), extOnly = FALSE)
}
\arguments{
\item{n0}{initial number of species, usually 1. Good parameter
to tweak if one is observing a low sample size when testing.}

\item{pp}{speciation rate. Must be constant.}

\item{qq}{extinction rate, similar to above.

Note: \code{pp} and \code{qq} must always be greater than 0}

\item{tMax}{ending time of simulation. Any species still living
after tMax is considered extant, and any species that would be generated
after \code{tMax} is not born.}

\item{nFinal}{an interval of acceptable number of species at the end of the
simulation. If not supplied, default is \code{c(0, Inf)}, so that any number
of species is accepted. If supplied, \code{BDSimConstant} will run until the
number of total species generated, or, if \code{extOnly = TRUE}, the number of
extant species at the end of the simulation, lies within the interval.}

\item{extOnly}{a boolean indicating whether \code{nFinal} should be taken as
the number of total or extant species during the simulation. If \code{TRUE},
\code{BDSimConstant} will run until the number of extant species lies within
the \code{nFinal} interval. If \code{FALSE}, as default, it will run until the
total number of species generated lies within that interval.}
}
\value{
a list of vectors, as follows

\describe{
\item{\code{TE}}{list of extinction times, with -0.01 as the time of
extinction for extant species.}

\item{\code{TS}}{list of speciation times, with tMax+0.01 as the time of
speciation for species that started the simulation.}

\item{\code{PAR}}{list of parents. Species that started the simulation have
NA, while species that were generated during the simulation have their
parent's number. Species are numbered as they are born.}

\item{\code{EXTANT}}{list of booleans representing whether each species is
extant.}}
}
\description{
\code{BDSimConstant} takes a rate, a maximum simulation time and an initial
number of species and returns the history of the clade originated from those
species using a birth-death process with constant rates equal to the given 
rates. It then generates the speciation and extinction times, parent and 
status information for the species in the simulation. Time runs from 
\code{tMax} to 0, to be consistent with the literature, though one can easily 
invert that by subtracting the results from \code{tMax}.
}
\examples{

# we can show a couple scenarios

###
# first, extinction 0

# initial number of species
n0 <- 1

# maximum simulation time
tMax <- 40

# speciation
p <- 0.1

# extinction
q <- 0

# run the simulation
sim <- BDSimConstant(n0, p, q, tMax, nFinal = c(2, Inf))

# we can plot the phylogeny to take a look
if (requireNamespace("ape", quietly = TRUE)) {
  phy <- MakePhylo(sim)
  ape::plot.phylo(phy)
}

###
# now let us try to turn extinction up a bit

# initial number of species
n0 <- 1

# maximum simulation time
tMax <- 40

# speciation
p <- 0.1

# extinction
q <- 0.04

# run the simulation
sim <- BDSimConstant(n0, p, q, tMax, nFinal = c(2, Inf))

# we can plot the phylogeny to take a look
if (requireNamespace("ape", quietly = TRUE)) {
  phy <- MakePhylo(sim)
  ape::plot.phylo(phy)
}

###
# we can also try a pure-death process

# initial number of species - note the high number, so we get an appreciable
# sample size
n0 <- 100

# maximum simulation time
tMax <- 40

# speciation
p <- 0

# extinction
q <- 0.02

# run the simulation
sim <- BDSimConstant(n0, p, q, tMax)

# of course in this case there are no phylogenies to plot

}
\author{
written by Bruno do Rosario Petrucci.
}
