% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSimGeneral.R
\name{BDSimGeneral}
\alias{BDSimGeneral}
\title{Returns information of a simulated clade for general speciation and
extinction rates}
\usage{
BDSimGeneral(N0, pp, qq, tmax, sshape = NULL, eshape = NULL)
}
\arguments{
\item{N0}{initial number of species, usually 1. Good param to
tweak if one is observing a low sample size when testing.}

\item{pp}{function to hold the speciation rate over time.
\code{BDSim} supplies this function with a \code{pp} ready to be used, so
that the only other information \code{BDSimGeneral} needs is a shape in case
the rate is to be age-dependent.}

\item{qq}{similar to above, but for extinction rate.}

\item{tmax}{ending time of simulation. Any species still living
after \code{tmax} is considered extant, and any species that would be
generated after \code{tmax} is not born.}

\item{sshape}{shape param for the Weibull distribution for
age-dependent speciation. Default is 0, where \code{pp} will be considered a
time-dependent exponential rate. For \code{sshape != NULL}, \code{pp} will
be considered a scale, and \code{rexp_var} will draw a Weibull distribution
instead.}

\item{eshape}{similar as above, but for extinction rate.}
}
\value{
a list of vectors, as follows

\describe{
\item{\code{TE}}{list of extinction times, with -0.01 as the time of
extinction for extant species.}

\item{\code{TS}}{list of speciation times, with tmax+0.01 as the time of
speciation for species that started the simulation.}

\item{\code{PAR}}{list of parents. Species that started the simulation have
NA, while species that were generated during the simulation have their
parent's number. Species are numbered as they are born.}

\item{\code{EXTANT}}{list of booleans representing whether a species is
extant.}}
}
\description{
\code{BDSimGeneral} takes an initial number of species, speciation and
extinction rates (either functions of time or of time and some
environmental variable), a maximum simulation time and possibly a shape for
age-dependent speciation and/or extinction.
}
\examples{

# REMEMBER TO INVERT TE AND TS IN SIMMEAN()

}
\author{
written by Bruno do Rosario Petrucci.
}
