% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{binner}
\alias{binner}
\title{Bin occurrences in geologic intervals}
\usage{
binner(x, bins)
}
\arguments{
\item{x}{The vector containing occurrence times for a given species.}

\item{bins}{A vector of time intervals corresponding to geological time ranges.}
}
\value{
A vector of occurrence counts for each interval.
}
\description{
Given a vector of fossil occurrences and time bins to represent geological
ranges, returns the occurrence counts in each bin.
}
\examples{

###
# first let us create some artificial occurrence data and check

# occurrence vector
x <- c(5.2, 4.9, 4.1, 3.2, 1, 0.2)

# bins vector
bins <- c(6, 5, 4, 3, 2, 1, 0)

# result
binnedSamp <- binner(x, bins)
binnedSamp

###
# it should work with any type of number in bins

# occurrence vector
x <- c(6.7, 5.03, 4.2, 3.4, 1.2, 0.4)

# bins vector
bins <- c(7.2, 6.1, 5.6, 4.3, 3.2, sqrt(2), 1, 0)

# result
binnedSamp <- binner(x, bins)
binnedSamp

###
# let us try with a real simulated species fossil record

# run the simulation
sim <- bd.sim(1, lambda = 0.1, mu = 0.05, tMax = 15)

# sample it
sampled <- sample.species(sim = sim, rho = 1, tMax = 15, S = 1)

# bins vector
bins <- c(15.1, 12.3, 10, 7.1, 5.8, 3.4, 2.2, 0)

# result
binnedsample <- binner(sampled, bins)
binnedsample

### 
# just one more

# run the simulation
sim <- bd.sim(1, lambda = function(t) {
  return(0.05 + 0.005*t)
}, mu = 0.05, tMax = 20)

# sample it
sampled <- sample.species(sim = sim, rho = 1, tMax = 20, S = 1)

# bins vector
bins <- c(15.1, 12.3, 10, 7.1, 5.8, 3.4, 2.2, 0)

# result
binnedsample <- binner(sampled, bins)
binnedsample

}
\author{
Matheus Januario and Bruno do Rosario Petrucci
}
