% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo.R, R/buddPhyloTreeHandling.R
\name{buddPhylo}
\alias{buddPhylo}
\alias{is.buddPhylo}
\alias{print.buddPhylo}
\alias{head.buddPhylo}
\alias{tail.buddPhylo}
\alias{summary.buddPhylo}
\alias{getParents.buddPhylo}
\alias{getChildren.buddPhylo}
\alias{getDescendants.buddPhylo}
\alias{getMRCA.buddPhylo}
\alias{is.monophyletic.buddPhylo}
\title{Details, generics, and methods for the \code{buddPhylo} class}
\usage{
is.buddPhylo(x)

\method{print}{buddPhylo}(x, ...)

\method{head}{buddPhylo}(x, subType = NULL, ...)

\method{tail}{buddPhylo}(x, subType = NULL, ...)

\method{summary}{buddPhylo}(object, ...)

getParents.buddPhylo(buddPhylo, focalLineage)

getChildren.buddPhylo(buddPhylo, focalLineage, returnInfo = TRUE)

getDescendants.buddPhylo(buddPhylo, focalLineage, onlyImmediates = FALSE)

getMRCA.buddPhylo(buddPhylo, nameList)

is.monophyletic.buddPhylo(buddPhylo, nameList, excludeSampAnc = TRUE)
}
\arguments{
\item{...}{Further arguments inherited from generics.}

\item{subType}{A character telling which "type" should be used to subset the
output, for instance only tips \code{subType="tip"} or only sampled ancestors
\code{subType="sampAnc"}. If set to \code(NULL) (default), it shows all elements.}

\item{buddPhylo, x, object}{Object of class "buddPhylo"}

\item{focalLineage}{A node/lineage to be used as reference for the tree operation.}

\item{onlyImmediates}{Logical indicating if only the most immediate
descendant (the two daughter) lineages of a given \code{focalLineage} should
be considered.}

\item{nameList}{A character vector with a list of \code{name} (i.e., not
"taxon") to use as a reference for the node set operation.}

\item{excludeSampAnc}{Logical indicating whether sampled ancestors should be
exlcuded from the output.}
}
\description{

}
\details{
\code{is.buddPhylo} A \code{buddPhylo} object must contain all 9
members: \code{"name}, \code{length}, \code{orientation}, \code{taxon},
\code{parent}, \code{type}, \code{y_coord}, \code{x_coord"}. And all of
these must have the correct length (i.e. \code{= (2*Ntip)-1)}.

\code{print.buddPhylo} The printing of a buddPhylo object is formatted into a more
straightforward and informative sequence manner. We provide details only for
the first few species, since otherwise this print could be overwhelming for
simulations with 10+ species.

\code{head.buddPhylo} Selects only a number of species from the
beginning of a \code{buddPhylo} object.

\code{tail.buddPhylo} Selects only a number of species from the end of a
\code{buddPhylo} object.

\code{summary.buddPhylo} Quantitative details on the \code{buddPhylo} object.
Prints the number of species, number of extant species, summary of durations
and speciation waiting times, in case there are more than one species.

\code{getParents.buddPhylo} Find all the parental lineages of a
\code{focalLineage} from a \code{buddPhylo$name} object, from first to
last ancestor. Note this does not use the information in the
\code{buddPhylo$taxon} column.

\code{getChildren.buddPhylo} returns the immediate descendant
(daughter) lineages of a given \code{focalLineage} in a \code{buddPhylo}
object (using the \code{name} column). The output is a character vector of
length two containing the \code{name} of each child, equivalent to
\code{getDescendants.buddPhylo(..., onlyImmediates = TRUE)}.

By default (\code{returnInfo = TRUE}), the function also annotates each
descendant based on its status. If \code{focalLineage} leads to a sampled
ancestor, one child is labeled accordingly and the other as a "lineage". If
it instead \code{focalLineage} leads to a speciation event, descendants are
labeled according to their \code{orientation}. Note that this function does
not use information from the \code{buddPhylo$taxon} column.

\code{getDescendants.buddPhylo} Find all the descendant lineages of a
\code{focalLineage} from a \code{buddPhylo$name} object, from most immediate
to most distal descendant. \code{getChildren.buddPhylo} returns only the
most immediate descendants (same effect as \code{getDescendants.buddPhylo(..., onlyImmediates = T)}). Note this does not use the information in the
\code{buddPhylo$taxon} column.

\code{getDescendants.buddPhylo} Find all the descendant lineages of a
\code{focalLineage} from a \code{buddPhylo$name} object, from most immediate
to most distal descendant. Note this does not use the information in the
\code{buddPhylo$taxon} column.

\code{getDescendants.buddPhylo} Find all the descendant lineages of a
\code{focalLineage} from a \code{buddPhylo$name} object, from most immediate
to most distal descendant. Note this does not use the information in the
\code{buddPhylo$taxon} column.
}
\author{
Matheus Januario
}
