% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{build.buddPhylo}
\alias{build.buddPhylo}
\title{Build a buddPhylo object}
\usage{
build.buddPhylo(newick, parsedTxt, taxonDiction = NULL, tolExtant = 1e-09)
}
\arguments{
\item{newick}{A character containing an annotated tree in newick
(parenthetic) format (e.g., as output from \code{parse.nexus}).}

\item{parsedTxt}{A data frame similar to the one generated by
\code{parse.newick.annotations}. It must contain node information on 
orientation, branch length, node name, and any other annotations associated 
with every node in a tree.}

\item{taxonDiction}{A data frame similar to the output of 
\code{extract.translate.block.nexus}, containing the node \code{id} and 
the \code{taxon} associated with it.}

\item{tolExtant}{Rounding error tolerance for assigning a tip as extant,
given the distance in time from the tip that is further from the root.
Default is \code{10E-10}.}
}
\value{
A \code{buddPhylo} object.
}
\description{
This function builds a \code{buddPhylo} object from the data contained in
a newick tree.
}
\author{
Matheus Januario and Bruno do Rosario Petrucci.
}
