% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.sim.R
\name{draw.sim}
\alias{draw.sim}
\title{Draw a sim object}
\usage{
draw.sim(sim, fossils = NULL, sort_by = "TS")
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation 
times, parent, and status information for each species in the simulation. 
See \code{?sim}.}

\item{fossils}{A \code{data.frame} containing the fossil occurrences of 
each lineage, as returned by the \code{sample.clade} function.}

\item{sort_by}{A single character or integer vector indicating how lineages 
should be sorted in the plot. If inputted as character (see example 3), it 
indicates which element in the \code{sim} object that should be used to sort 
lineages in the plot. If inputted as a vector of integers, it directly 
specifies the order in which lineages should be drawn (from the bottom 
(i.e. integer = 1)) to the upper side (integer = length of the \code{sim} 
elements) of the figure). Default value of this parameter is "TS".}
}
\value{
A plot of the simulation in the graphics window. If the 
\code{fossils} data.frame is inputted, its format will dictate how fossil
occurrences will be plotted. If \code{fossils} has a \code{SampT} column
(i.e. the occurrence times are exact), fossil occurrences are assigned as dots. 
If \code{fossils} has columns \code{MaxT} and \code{MinT} (i.e. the early and 
late stage bounds associated with each occurrence), fossil occurrences are 
represented as slightly jittered, semitransparent bars indicating the early 
and late bounds of each fossil occurrence.
}
\description{
Draws a paleobuddy simulation (a \code{sim} object - please see 
\code{?sim}) in the graphics window. Allows for the assignment of speciation
or sampling events, and further customization.
}
\examples{

###
# start simple

# maximum simulation time
tMax <- 10 

# run birth-death simulation
sim <- bd.sim(n0 = 1, lambda = 0.6, mu = 0.55, tMax = tMax, 
              nFinal = c(10,20)) 

# simulate data resulting from a fossilization process
# with exact occurrence times
fdt <- sample.clade(sim = sim, rho = 4, tMax = tMax, returnTrue = TRUE)

# draw it
draw.sim(sim, fossils = fdt, sort_by = "PAR")

###
# try with fossil ranges

# maximum simulation time
tMax <- 10 

# run birth-death simulation
sim <- bd.sim(n0 = 1, lambda = 0.6, mu = 0.55, tMax = tMax, 
              nFinal = c(10,20)) 

# simulate data resulting from a fossilization process
# with fossil occurrence time ranges

# create time bins randomly
bins <- c(tMax, 0, runif(n = rpois(1, lambda = 6), min = 0, max = tMax))

# simulate fossil sampling
fdt <- sample.clade(sim = sim, rho = 2, tMax = tMax, 
                    returnTrue = FALSE, bins = bins)

# draw it, sorting lineages by their parent
draw.sim(sim, fossils = fdt, sort_by = "PAR")

# adding the bounds of the simulated bins
abline(v = bins, lty = 2, col = "red", lwd = 0.5)

###
# we can control how to sort displayed species exactly

# maximum simulation time
tMax <- 10 

# run birth-death simulations
sim <- bd.sim(n0 = 1, lambda = 0.6, mu = 0.55, tMax = tMax, 
              nFinal = c(10,20)) 
 
# simulate fossil sampling
fdt <- sample.clade(sim = sim, rho = 4, tMax = tMax, returnTrue = TRUE)
 
# draw it with random sorting
draw.sim(sim, fossils = fdt, sort_by = sample(1:length(sim$TE)))

}
\author{
Matheus Januario
}
