% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{extract.translate.block.nexus}
\alias{extract.translate.block.nexus}
\title{Extract the "translation" block in a nexus file}
\usage{
extract.translate.block.nexus(text = NULL, file = NULL)
}
\arguments{
\item{file}{A character containing a nexus file name.}
}
\value{
A data frame containing the number \code{id} and the \code{taxon}
associated with it.
}
\description{
This function only the tip translations (numbers to characters) in NEXUS file.
}
\author{
Matheus Januario
}
