% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{extract.translate.block.nexus}
\alias{extract.translate.block.nexus}
\title{Extract the "translation" block in a nexus file}
\usage{
extract.translate.block.nexus(text = NULL, file = NULL)
}
\arguments{
\item{text}{A string with the tree in a Newick format. Optional, but at least
one of \code{text} or \code{file} must be not \code{NULL}.}

\item{file}{A character containing a Nexus filename. Optional, but at least
one of \code{text} or \code{file} must be not \code{NULL}.}
}
\value{
A data frame containing the number \code{id} and the \code{taxon}
associated with it.
}
\description{
This function extracts the tip translation block (the block relating numbers
to taxon names) in a Nexus file.
}
\author{
Matheus Januario and Bruno do Rosario Petrucci.
}
