% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.buddPhylo.R
\name{make.buddPhylo}
\alias{make.buddPhylo}
\title{Generate a budding phylogenetic tree from simulations}
\usage{
make.buddPhylo(sim, fossils = NULL, returnTrueExt = TRUE)
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation 
times, parent, and status information for each species in the simulation. 
See \code{?sim}.}

\item{fossils}{A data frame with a \code{"Species"} column and a
\code{SampT} column, usually an output of the \code{sample.clade}
function. Species names must contain only one number each, corresponding
to the order of the \code{sim} vectors. Note that we require it to have a
\code{SampT} column, i.e. fossils must have an exact age. This assumption
might be relaxed in the future.}

\item{returnTrueExt}{A logical indicating whether to include in tree the
tips representing the true extinction time of extinct species. If set to
\code{TRUE} (default), the returned tree will include those tips. If 
\code{FALSE}, they will be dropped and instead the last sampled fossil of
a given species will be the last sampled tip of that species. Note that if
a species was not sampled it will then not appear in the tree. If no fossils
have been added to the tree with the parameter \code{fossils}, this will
have the equivalent effect as the ape function \code{drop.fossil}, returning
an ultrametric tree.}
}
\value{
... TODO
}
\description{
\code{make.buddPhylo} generates a budding phylogenetic tree
from a birth-death simulation (e.g. an output from \code{bd.sim}) and 
(optionally) a fossil record (e.g. an output from \code{sample.clade}).
}
\details{
.... TODO
}
\author{
Bruno do Rosario Petrucci
}
