% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.phylo.R
\name{make.phylo}
\alias{make.phylo}
\title{Phylogeny generating}
\usage{
make.phylo(sim, fossils = NULL)
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation times,
parent, and status information for each species in the simulation. See 
\code{?sim}.}

\item{fossils}{A data frame with a \code{"Species"} column and a
\code{SampT} column, usually an output of the \code{sample.clade}
function. Species names must contain only one number each, corresponding
to the order of the \code{sim} vectors.}
}
\value{
A \code{phylo} object from the APE package. Tip labels are numbered
following the order of species in the \code{sim} object. If fossil occurrence
data was supplied, the tree will include fossil occurrences as tips with branch
length \code{0}, bifurcating at its sampling time from the corresponding
species' edge (i.e. a sampled ancestor tree, sensu Heath et al, (2014)). Note
that to obtain a true SA tree, one must perform the last step of deleting tips
that are not either extant or fossil occurrences (i.e. the tips at true time of
extinction).

Note this package does not depend on APE (Paradis et al, 2004) since it is 
never used inside its functions, but it is suggested since one might want to
manipulate the phylogenies generated by this function.
}
\description{
Generates a phylogeny from a \code{sim} object containing speciation and 
extinction times, parent and status information (see \code{?sim}). Returns a
\code{phylo} object containing information on the phylogeny, following an 
"evolutionary Hennigian" (sensu Ezard et al 2011) format. Takes an optional
argument encoding fossil occurrences to return a Sampled Ancestor tree (see
references), with fossil occurrences as tips with branch length \code{0}.
Returns \code{NA} and sends a warning if the simulation has only one lineage or
if more than one species has \code{NA} as parent (i.e. there is no single 
common ancestor in the simulation). In the latter case, please use 
\code{find.lineages} first.
}
\examples{

###
# generating a phylogeny using constant rates
sim <- bd.sim(n0 = 1, lambda = 0.2, mu = 0.05, tMax = 10, 
             nExtant = c(2, Inf))

# make the phylogeny
phy <- make.phylo(sim)

par(mfrow = c(1, 2))
# plot it
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  
  # we can also plot only the molecular phylogeny
  ape::plot.phylo(ape::drop.fossil(phy))
}

###
# this works for sim generated with any of the scenarios in bd.sim, of course
sim <- bd.sim(n0 = 1, lambda = function(t) 0.05 + 0.01*t, 
             mu = function(t) 0.03 + 0.015*t, tMax = 10, 
             nExtant = c(2, Inf))

# make the phylogeny
phy <- make.phylo(sim)

### 
# Let us try out the fossils argument to generate a Sample Ancestor tree

# simulate a simple birth-death process
sim <- bd.sim(n0 = 1, lambda = 0.2, mu = 0.05, tMax = 10, 
              nExtant = c(2, Inf))

# make the traditional phylogeny
phy <- make.phylo(sim)

# sample fossils
fossils <- sample.clade(sim, 0.1, 10)

# make the sampled ancestor tree
saTree <- make.phylo(sim, fossils)

par(mfrow = c(1, 2))
# plot them
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy, main = "Phylogenetic tree")
  ape::axisPhylo()
  
  ape::plot.phylo(saTree, main = "Sampled Ancestor tree")
  ape::axisPhylo()
}

}
\references{
Ezard, T. H., Pearson, P. N., Aze, T., & Purvis, A. (2012). The meaning of 
birth and death (in macroevolutionary birth-death models). Biology letters, 
8(1), 139-142.

Paradis, E., Claude, J., Strimmer, & K. (2004). APE: Analyses of Phylogenetics
and Evolution in R language. Bioinformatics, 20(2), 289-290.

Heath, T. A., Huelsenbeck, J. P., & Stadler, T. (2014). The fossilized 
birth–death process for coherent calibration of divergence-time estimates. 
Proceedings of the National Academy of Sciences, 111(29), E2957-E2966.
}
\author{
Matheus Januario and Bruno do Rosario Petrucci
}
