% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.phylo.R
\name{make.phylo}
\alias{make.phylo}
\title{Converts a paleobuddy simulation into a phylogeny}
\usage{
make.phylo(sim)
}
\arguments{
\item{sim}{a simulation from the \code{bd.sim} function.}
}
\value{
A \code{phylo} object.
}
\description{
\code{make.phylo} generates a \code{phylo} object using a simulation from the
\code{bd.sim} function. The phylogeny follows a "Hennigian" (sensu Ezard et
al 2011) format. If the simulation has only one lineage, the function
returns \code{NA} as there is no phylogeny for a simulation with only one
lineage.
}
\examples{

# generating a phylogeny using constant rates
sim <- bd.sim(n0 = 1, pp = 0.2, qq = 0.05, tMax = 10, 
             nFinal = c(2, Inf), extOnly = TRUE)

# make the phylogeny
phy <- make.phylo(sim)

par(mfrow = c(1, 2))
# plot it
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  
  # we can also plot only the molecular phylogeny
  ape::plot.phylo(ape::drop.fossil(phy))
}

# this works for sim generated with any of the scenarios in bd.sim, of course
sim <- bd.sim(n0 = 1, pp = function(t) 0.05 + 0.01*t, 
             qq = function(t) 0.03 + 0.015*t, tMax = 10, 
             nFinal = c(2, Inf), extOnly = TRUE)

# make the phylogeny
phy <- make.phylo(sim)

par(mfrow = c(1, 2))
# plot it
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  
  # we can also plot only the molecular phylogeny
  ape::plot.phylo(ape::drop.fossil(phy))
}

}
\author{
Function written by Matheus Januario. 

Reference: Ezard, T. H.,
Pearson, P. N., Aze, T., & Purvis, A. (2012). The meaning of birth and death
(in macroevolutionary birth-death models). Biology letters, 8(1), 139-142.
}
