% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.phylo.R
\name{make.phylo}
\alias{make.phylo}
\title{Phylogeny generating}
\usage{
make.phylo(sim)
}
\arguments{
\item{sim}{A \code{sim} object, containing extinction times, speciation times,
parent, and status information for each species in the simulation. See 
\code{?sim}.}
}
\value{
A \code{phylo} object from the APE package. Tip labels are numbered
following the order of species in the \code{sim} object.

Note this package does not
depend on APE since it is never used inside its functions, but it is suggested
since one might want to manipulate the phylogenies generated by this function.
}
\description{
Generates a phylogeny from a \code{sim} object containing speciation and 
extinction times, parent and status information, usually returned by 
\code{bd.sim} or similar. Returns a \code{phylo} object containing information 
on the phylogeny, following a "Hennigian" (sensu Ezard et al 2011) format. 
Returns \code{NA} and sends a warning if the simulation has only one lineage or
if more than one species has \code{NA} as parent (i.e. there is no single 
common ancestor in the simulation). In the latter case, please use 
\code{find.lineages} first.
}
\examples{

###
# generating a phylogeny using constant rates
sim <- bd.sim(n0 = 1, lambda = 0.2, mu = 0.05, tMax = 10, 
             nFinal = c(2, Inf), extOnly = TRUE)

# make the phylogeny
phy <- make.phylo(sim)

par(mfrow = c(1, 2))
# plot it
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  
  # we can also plot only the molecular phylogeny
  ape::plot.phylo(ape::drop.fossil(phy))
}

###
# this works for sim generated with any of the scenarios in bd.sim, of course
sim <- bd.sim(n0 = 1, lambda = function(t) 0.05 + 0.01*t, 
             mu = function(t) 0.03 + 0.015*t, tMax = 10, 
             nFinal = c(2, Inf), extOnly = TRUE)

# make the phylogeny
phy <- make.phylo(sim)

par(mfrow = c(1, 2))
# plot it
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  
  # we can also plot only the molecular phylogeny
  ape::plot.phylo(ape::drop.fossil(phy))
}

}
\references{
Ezard, T. H., Pearson, P. N., Aze, T., & Purvis, A. (2012). The meaning of 
birth and death (in macroevolutionary birth-death models). Biology letters, 
8(1), 139-142.
}
\author{
Matheus Januario.
}
