% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePhylo.R
\name{MakePhylo}
\alias{MakePhylo}
\title{Converts a paleobuddy simulation into a phylogeny}
\usage{
MakePhylo(sim)
}
\arguments{
\item{sim}{a simulation from the \code{BDSim} function.}
}
\value{
A \code{phylo} object.
}
\description{
\code{MakePhylo} generates a \code{phylo} object using a simulation from the
\code{BDSim} function. The phylogeny follows a "Hennigian" (sensu Ezard et
al 2011) format. If the simulation has only one lineage, the function
returns \code{NA} as there is no phylogeny for a simulation with only one
lineage.
}
\examples{

# generating a phylogeny using constant rates
sim <- BDSim(n0 = 1, pp = 0.2, qq = 0.05, tMax = 10)

# in case first simulation has only one species
while (length(sim$TE) < 2) {
  sim <- BDSim(n0 = 1, pp = 0.2, qq = 0.05, tMax = 10)
}

phy <- MakePhylo(sim)

# we need ape to plot it
if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  
  # we can also plot only the molecular phylogeny
  ape::plot.phylo(ape::drop.fossil(phy))
}

# this works for sim generated with any of the scenarios in BDSim, of course
sim <- BDSim(n0 = 1, pp = function(t) 0.12 + 0.01*t, qq = 10, 
             tMax = 10, qShape = 1.3)

# in case first simulation has only one species
while(length(sim$TE) < 2) { 
  sim <- BDSim(n0 = 1, pp = function(t) 0.12 + 0.01*t, qq = 10, 
               tMax = 10, qShape = 1.3)
}
phy <- MakePhylo(sim)

if (requireNamespace("ape", quietly = TRUE)) {
  ape::plot.phylo(phy)
  ape::plot.phylo(ape::drop.fossil(phy))
}
}
\author{
Function written by Matheus Januario. 

Reference: Ezard, T. H.,
Pearson, P. N., Aze, T., & Purvis, A. (2012). The meaning of birth and death
(in macroevolutionary birth-death models). Biology letters, 8(1), 139-142.
}
