% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleobuddy.R
\docType{package}
\name{paleobuddy}
\alias{paleobuddy}
\title{paleobuddy: Simulating diversification dynamics with flexible scenarios}
\description{
Here is where I will write the help section for the package
}
\section{Help needed}{
NA
}

