% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleobuddy.R
\docType{package}
\name{paleobuddy}
\alias{paleobuddy}
\title{paleobuddy}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

