% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{parse.newick.annotations}
\alias{parse.newick.annotations}
\title{Parse and organize annotations from a newick tree}
\usage{
parse.newick.annotations(newick, taxonDiction)
}
\arguments{
\item{newick}{A character containing an annotated tree in newick
(parenthetic) format (e.g., as output from \code{parse.nexus}).}

\item{taxonDiction}{A data frame similar to the output of 
\code{extract.translate.block.nexus}, containing the node \code{id} and 
the \code{taxon} associated with it.}
}
\value{
A data frame containing node information on lineage (taxon +
first/last occurrence), taxon (taxonomic name associated with the branch),
orientation, branch length, node name, and any other annotations
associated with every node in a tree. Note the function does not return a
\code{buddPhylo} object.
}
\description{
This function finds and sorts any node-specific annotation in a nexus file.
}
\author{
Matheus Januario
}
