% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{parse.nexus}
\alias{parse.nexus}
\title{Subset and parse a nexus file}
\usage{
parse.nexus(text = NULL, file = NULL, treeNumbers = NULL)
}
\arguments{
\item{text}{A string with the tree in a Newick format. Optional, but at least
one of \code{text} or \code{file} must be not \code{NULL}.}

\item{file}{A character containing a Nexus filename. Optional, but at least
one of \code{text} or \code{file} must be not \code{NULL}.}

\item{treeNumbers}{A numeric vector containing which trees should be read.
If \code{NULL} (default), the function reads all trees within \code{file}.}
}
\value{
A character with the parenthetic format within the nexus file.
If more than one tree is read, a list of characters is returned in the
order given by \code{treeNumbers}.
}
\description{
This function parses one or several budding phylogenetic trees in a Nexus 
file.
}
\author{
Matheus Januario and Bruno do Rosario Petrucci.
}
\keyword{internal}
