% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{read.nexus.buddPhylo}
\alias{read.nexus.buddPhylo}
\title{Read a buddPhylo tree from a file in Nexus format}
\usage{
read.nexus.buddPhylo(file, treeNumbers = NULL, timeFromRoot = FALSE)
}
\arguments{
\item{file}{A character containing a Nexus filename. Optional, but at least
one of \code{text} or \code{file} must be not \code{NULL}.}

\item{treeNumbers}{A numeric vector containing which trees should be read.
If \code{NULL} (default), the function reads all trees within \code{file}.}

\item{timeFromRoot}{A logical indicating whether time should be expressed as
time since the root (\code{TRUE}). If set to \code{FALSE} (default), time is 
expressed as millions of years ago (Mya), taking the tip extant tip that is 
further from the root as a reference for "present".}
}
\value{
A \code{buddPhylo} with the time adjusted accordingly.
}
\description{
This function reads one or several trees in a Nexus file.
}
\details{
Note that internal nodes must have labels.
}
\examples{

###
# a small annotated tree in Nexus format ships with the package
f <- system.file("extdata", "example.nex", package = "paleobuddy")

# read it
budd <- read.nexus.buddPhylo(f)

# read only specific trees from a multi-tree file
budd2 <- read.nexus.buddPhylo(f, treeNumbers = 1)

# time can be expressed from the root instead of Mya
budd_root <- read.nexus.buddPhylo(f, timeFromRoot = TRUE)

}
\author{
Matheus Januario and Bruno do Rosario Petrucci.
}
