% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_reader.R
\name{read.trees.buddPhylo}
\alias{read.trees.buddPhylo}
\title{Read a buddPhylo tree from a file in Newick format}
\usage{
read.trees.buddPhylo(
  file,
  text = NULL,
  treeNumbers = NULL,
  timeFromRoot = FALSE
)
}
\arguments{
\item{file}{A character containing a Newick file name.}

\item{text}{A string with one or several trees in Newick format, separated
by newlines (\code{\\n}). \code{NULL} by default. If not \code{NULL}, the 
argument \code{file} is ignored.}

\item{treeNumbers}{A numeric vector containing which trees should be read.
If \code{NULL} (default), the function reads all trees within \code{file}.}

\item{timeFromRoot}{A logical indicating whether time should be expressed as
time since the root (\code{TRUE}). If set to \code{FALSE} (default), time is 
expressed as millions of years ago (Mya), taking the tip extant tip that is 
further from the root as a reference for "present".}
}
\value{
A \code{buddPhylo} with the time adjusted accordingly.
}
\description{
This function reads one or several trees in Newick file.
}
\details{
Note that internal nodes must have labels.
}
\examples{

###
# buddPhylo trees are Newick trees with [&...] annotations on each node,
# minimally an orientation ("ancestor", or "descendant")
nw <- paste0("((t1_last[&orientation=ancestor,range=t1]:1,",
             "t1_first[&orientation=ancestor]:0)",
             "2[&orientation=ancestor]:1,",
             "t2_first[&orientation=descendant]:2)",
             "1[&orientation=ancestor]:0;")

# read it
budd <- read.trees.buddPhylo(text = nw)

# note taxon strips the _first/_last occurrence suffix
budd[, c("name", "lineage", "taxon", "type")]

# several trees can be read at once, one per element
budds <- read.trees.buddPhylo(text = c(nw, nw))

# and you can pick which ones to read
just_second <- read.trees.buddPhylo(text = c(nw, nw), treeNumbers = 2)

}
\author{
Bruno do Rosario Petrucci.
}
