% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp.var.R
\name{rexp.var}
\alias{rexp.var}
\title{General rate exponential and Weibull waiting times}
\usage{
rexp.var(
  n = 1,
  lambda,
  now = 0,
  tMax = Inf,
  shape = NULL,
  TS = NULL,
  fast = FALSE
)
}
\arguments{
\item{n}{The number of waiting times to return. The default is 1, but we
allow for a higher \code{n} to be consistent with the \code{rexp} function.}

\item{lambda}{The rate parameter for the exponential distribution. If
\code{shape} is not \code{NULL}, \code{lambda} is a scale for a Weibull
distribution. In both cases we allow for any time-varying function. If one
wants to use a constant rate, please use 
\code{lambda <- Vectorize(function(t) c)}, or \code{rexp}.}

\item{now}{The current time. Needed so that we consider only the interval
between the current time and the maximum time for the time-varying rate.
Notice this does means the waiting time is \code{>= now}, so we also
subtract \code{now} from the result before returning. The default is \code{0}.}

\item{tMax}{The simulation ending time. If the waiting time would be too
high, we return \code{tMax + 0.01} to signify the event never happens, if
\code{FAST == TRUE}. Otherwise we return the true waiting time. By default,
\code{tMax} will be \code{Inf}, but if \code{FAST == TRUE} one must supply
a finite value.}

\item{shape}{The shape of a weibull distribution. If not \code{NULL}, the 
distribution is taken to be a weibull. Otherwise, it is considered an
exponential.

Note: Time-varying shape is implemented, so one could have \code{pShape} or
\code{qShape} be a function of time. It is not thoroughly tested, however, so 
it may be prudent to wait for a future release where this feature is well 
established.

Note: Time-varying shape is implemented, so one could have \code{shape} be a
function of time. It is not thoroughly tested, however, so it may be prudent to
wait for a future release where this feature is well established.}

\item{TS}{If shape is given, there must be a \code{TS} parameter to account
for the scaling between simulation and species time. Supplying one without the
other leads to an error. The default is \code{0}.}

\item{fast}{If set to \code{FALSE}, waiting times larger than \code{tMax} will
not be thrown away. This argument is needed so one can test the function
without bias.}
}
\value{
A vector of waiting times for the exponential or weibull distribution
with the given parameters.
}
\description{
Generates a waiting time following an exponential or Weibull distribution with
constant or varying rates. Allows for an optional shape parameter, in which 
case \code{lambda} will be taken as a Weibull scale. Requires information on
the current time, to consider only rates starting from then, and the speciation
time, optionally, in case shape is provided and the process is age-dependent.
Allows for customization on the possibility of throwing away a waiting time
higher than \code{tMax}, and therefore takes that time as a parameter.
}
\examples{

# let us start by checking a simple exponential variable

# rate
lambda <- 0.1

# current time
now <- 0

# maximum time to check
tMax <- 40

# get waiting time
t <- rexp.var(n = 1, lambda, now, tMax)
t

###
# another simple exponential

# rate
lambda <- 0.5

# current time
now <- 0

# maximum time to check
tMax <- 40

# find the waiting time
t <- rexp.var(n = 1, lambda, now, tMax)
t

###
# now let us try a linear function for the rate

# rate
lambda <- function(t) {
  return(0.01*t + 0.1)
}

# current time
now <- 0

# maximum time to check
tMax <- 40

# find the waiting time
t <- rexp.var(n = 1, lambda, now, tMax)
t

###
# what if lambda is exponential?

# rate
lambda <- function(t) {
  return(0.01 * exp(0.1*t) + 0.02)
}

# current time
now <- 0

# maximum time to check
tMax <- 40

# find the waiting time
t <- rexp.var(n = 1, lambda, now, tMax)
t

###
# rexp.var also works for a weibull

# scale
lambda <- 2

# shape
shape <- 1

# current time
now <- 0

# maximum time to check
tMax <- 40

# speciation time
TS <- 0

# find the vector of waiting time
t <- rexp.var(n = 1, lambda, now, tMax,
              shape = shape, TS = TS)
t

###
# when shape = 1, the Weibull is an exponential, we could do better

# scale
lambda <- 10

# shape
shape <- 2

# current time
now <- 0

# maximum time to check
tMax <- 40

# speciation time
TS <- 0

# find the vector of waiting times - it doesn't need to be just one
t <- rexp.var(n = 5, lambda, now, tMax,
              shape = shape, TS = TS)
t

###
# shape can be less than one, of course

# scale
lambda <- 10

# shape
shape <- 0.5

# current time
now <- 3

# maximum time to check
tMax <- 40

# speciation time - it will be greater than 0 frequently during a simulation,
# as it is used to represent where in the species life we currently are and
# rescale accordingly
TS <- 2.5

# find the vector of waiting times
t <- rexp.var(n = 3, lambda, now, tMax,
              shape = shape, TS = TS)
t

###
# both lambda and shape can be time varying for a Weibull, but since we have
# only been able to test time-varying lambda effectively, we do not recommend
# using shape as a function

# scale
lambda <- function(t) {
  return(0.25*t + 5)
}

# shape
shape <- 3

# current time
now <- 0

# maximum time to check
tMax <- 40

# speciation time
TS <- 0

# find the vector of waiting times - it doesn't need to be just one
t <- rexp.var(n = 5, lambda, now, tMax,
              shape = shape, TS = TS)
t

###
# lambda can be any function of time, remember

# scale
lambda <- function(t) {
  return(0.2*exp(0.1*t) + 5)
}

# shape
shape <- 3

# current time
now <- 0

# maximum time to check
tMax <- 40

# speciation time
TS <- 0

# find the vector of waiting times - it doesn't need to be just one
t <- rexp.var(n = 2, lambda, now, tMax,
              shape = shape, TS = TS)
t

}
\author{
Bruno do Rosario Petrucci.
}
