% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_utilities.R
\name{subtrees.buddPhylo}
\alias{subtrees.buddPhylo}
\title{All subtrees of a budding phylogenetic tree}
\usage{
subtrees.buddPhylo(buddPhylo)
}
\arguments{
\item{buddPhylo}{Object of class "buddPhylo".}
}
\value{
A list of \code{buddPhylo} objects, one per internal node of 
\code{buddPhylo}, each rooted at that node.
}
\description{
This function returns a list of all the subtrees of a budding phylogenetic
tree, i.e. a list of \code{buddPhylo} objects.
}
\author{
Bruno do Rosario Petrucci.
}
