% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buddPhylo_utilities.R
\name{subtrees.buddPhylo}
\alias{subtrees.buddPhylo}
\title{List the monophyletic clades in a budding phylogeny}
\usage{
subtrees.buddPhylo(buddPhylo, returnInfo = TRUE)
}
\arguments{
\item{buddPhylo}{Object of class "buddPhylo"}

\item{returnInfo}{Whether to return the MRCA (if the defining node of the
clade is a sampled-ancestor node) or orientation (if the defining node of
the clade is a speciation node) for each clade. Defaults to \code{TRUE}.}
}
\value{
A list of length equal to the number of nodes, with each element
being a vector of strings (listing the tips/SAs that are descended
from each node). If \code{returnInfo = TRUE}, each element contains two 
vectors of strings: either the SA and lineage, or the ancestor clade and
the descendant clade (similar to \code{getChildren.buddPhylo}).
}
\description{
Get a list of monophyletic clades in this budding phylogeny.
}
\author{
Bruno do Rosario Petrucci
}
